trivial -> teaching completed Fixes #70
