+++ docs/_sources/opt.single_objective.trivial.max_ones_problem.rst.txt
@@ -1,61 +1,61 @@
-opt.single\_objective.trivial.max\_ones\_problem package
+opt.single\_objective.teaching.max\_ones\_problem package
 ========================================================
 
 Submodules
 ----------
 
-opt.single\_objective.trivial.max\_ones\_problem.command\_line module
+opt.single\_objective.teaching.max\_ones\_problem.command\_line module
 ---------------------------------------------------------------------
 
-.. automodule:: opt.single_objective.trivial.max_ones_problem.command_line
+.. automodule:: opt.single_objective.teaching.max_ones_problem.command_line
    :members:
    :undoc-members:
    :show-inheritance:
 
-opt.single\_objective.trivial.max\_ones\_problem.max\_ones\_problem module
+opt.single\_objective.teaching.max\_ones\_problem.max\_ones\_problem module
 --------------------------------------------------------------------------
 
-.. automodule:: opt.single_objective.trivial.max_ones_problem.max_ones_problem
+.. automodule:: opt.single_objective.teaching.max_ones_problem.max_ones_problem
    :members:
    :undoc-members:
    :show-inheritance:
 
-opt.single\_objective.trivial.max\_ones\_problem.max\_ones\_problem\_binary\_bit\_array\_solution module
+opt.single\_objective.teaching.max\_ones\_problem.max\_ones\_problem\_binary\_bit\_array\_solution module
 --------------------------------------------------------------------------------------------------------
 
-.. automodule:: opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution
+.. automodule:: opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution
    :members:
    :undoc-members:
    :show-inheritance:
 
-opt.single\_objective.trivial.max\_ones\_problem.max\_ones\_problem\_binary\_bit\_array\_solution\_vns\_support module
+opt.single\_objective.teaching.max\_ones\_problem.max\_ones\_problem\_binary\_bit\_array\_solution\_vns\_support module
 ----------------------------------------------------------------------------------------------------------------------
 
-.. automodule:: opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support
+.. automodule:: opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support
    :members:
    :undoc-members:
    :show-inheritance:
 
-opt.single\_objective.trivial.max\_ones\_problem.max\_ones\_problem\_binary\_int\_solution module
+opt.single\_objective.teaching.max\_ones\_problem.max\_ones\_problem\_binary\_int\_solution module
 -------------------------------------------------------------------------------------------------
 
-.. automodule:: opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution
+.. automodule:: opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution
    :members:
    :undoc-members:
    :show-inheritance:
 
-opt.single\_objective.trivial.max\_ones\_problem.max\_ones\_problem\_binary\_int\_solution\_vns\_support module
+opt.single\_objective.teaching.max\_ones\_problem.max\_ones\_problem\_binary\_int\_solution\_vns\_support module
 ---------------------------------------------------------------------------------------------------------------
 
-.. automodule:: opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support
+.. automodule:: opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support
    :members:
    :undoc-members:
    :show-inheritance:
 
-opt.single\_objective.trivial.max\_ones\_problem.solver module
+opt.single\_objective.teaching.max\_ones\_problem.solver module
 --------------------------------------------------------------
 
-.. automodule:: opt.single_objective.trivial.max_ones_problem.solver
+.. automodule:: opt.single_objective.teaching.max_ones_problem.solver
    :members:
    :undoc-members:
    :show-inheritance:
@@ -63,7 +63,7 @@ opt.single\_objective.trivial.max\_ones\_problem.solver module
 Module contents
 ---------------
 
-.. automodule:: opt.single_objective.trivial.max_ones_problem
+.. automodule:: opt.single_objective.teaching.max_ones_problem
    :members:
    :undoc-members:
    :show-inheritance:
