+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({"docnames": ["changelog", "contribution", "how_to_guides", "implemented_algorithms", "implemented_algorithms_exact", "implemented_algorithms_metaheuristic", "implemented_algorithms_total_enumeration", "implemented_algorithms_vns", "index", "modules", "opt", "opt.multi_objective", "opt.single_objective", "opt.single_objective.combinatorial", "opt.single_objective.combinatorial.traveling_salesperson_problem", "opt.single_objective.global", "opt.single_objective.trivial", "opt.single_objective.trivial.max_ones_problem", "problems_to_be_solved", "problems_to_be_solved_max_ones", "tutorials", "uo", "uo.algorithm", "uo.algorithm.exact", "uo.algorithm.exact.total_enumeration", "uo.algorithm.metaheuristic", "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic", "uo.algorithm.metaheuristic.genetic_algorithm", "uo.algorithm.metaheuristic.variable_neighborhood_search", "uo.target_problem", "uo.target_solution", "uo.utils"], "filenames": ["changelog.rst", "contribution.rst", "how_to_guides.rst", "implemented_algorithms.rst", "implemented_algorithms_exact.rst", "implemented_algorithms_metaheuristic.rst", "implemented_algorithms_total_enumeration.rst", "implemented_algorithms_vns.rst", "index.rst", "modules.rst", "opt.rst", "opt.multi_objective.rst", "opt.single_objective.rst", "opt.single_objective.combinatorial.rst", "opt.single_objective.combinatorial.traveling_salesperson_problem.rst", "opt.single_objective.global.rst", "opt.single_objective.trivial.rst", "opt.single_objective.trivial.max_ones_problem.rst", "problems_to_be_solved.rst", "problems_to_be_solved_max_ones.rst", "tutorials.rst", "uo.rst", "uo.algorithm.rst", "uo.algorithm.exact.rst", "uo.algorithm.exact.total_enumeration.rst", "uo.algorithm.metaheuristic.rst", "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic.rst", "uo.algorithm.metaheuristic.genetic_algorithm.rst", "uo.algorithm.metaheuristic.variable_neighborhood_search.rst", "uo.target_problem.rst", "uo.target_solution.rst", "uo.utils.rst"], "titles": ["Changelog", "How to Contribute", "How-to Guides", "Implemented algorithms", "Implemented exact algorithms", "Implemented metaheuristic algorithms", "Total Enumeration", "Variable Neighborhood Search", "Universal optimizer - Optimization algorithms library", "Library", "opt package", "opt.multi_objective package", "opt.single_objective package", "opt.single_objective.combinatorial package", "opt.single_objective.combinatorial.traveling_salesperson_problem package", "opt.single_objective.global package", "opt.single_objective.trivial package", "opt.single_objective.trivial.max_ones_problem package", "Problems to be solved", "Max Ones Problem", "Tutorials", "uo package", "uo.algorithm package", "uo.algorithm.exact package", "uo.algorithm.exact.total_enumeration package", "uo.algorithm.metaheuristic package", "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic package", "uo.algorithm.metaheuristic.genetic_algorithm package", "uo.algorithm.metaheuristic.variable_neighborhood_search package", "uo.target_problem package", "uo.target_solution package", "uo.utils package"], "terms": {"version": [0, 2], "0": [0, 17, 22, 24, 25, 28, 29, 30], "1": [0, 7, 17, 24, 25, 28, 30], "date": 0, "04": 0, "06": 0, "2023": 0, "descript": 0, "initi": [0, 1, 2, 7, 17, 22, 24, 28, 30], "setup": 0, "i": [0, 1, 2, 7, 8, 17, 19, 22, 24, 28, 30], "done": 0, "librari": [0, 1, 2], "structur": [0, 1, 3, 5, 8, 17, 25], "set": [0, 7, 17, 24, 29, 30], "2": [0, 7, 30], "26": 0, "metaheurist": [0, 1, 3, 7, 8, 21, 22], "algorithm": [0, 1, 9, 17, 21, 29], "vn": [0, 2, 7, 17, 24, 28], "ad": 0, "problem": [0, 1, 2, 7, 8, 17, 22, 24, 28, 29, 30], "maxon": 0, "4": 0, "01": 0, "07": 0, "document": [0, 1, 2, 7, 8], "support": [0, 7, 17], "should": [1, 22, 28, 31], "deriv": 1, "from": [1, 2, 7, 17, 28, 29, 30], "specifi": 1, "class": [1, 2, 7, 17, 19, 22, 24, 25, 28, 29, 30], "implement": [1, 8, 19, 28], "optim": [1, 7, 17, 22, 24, 25, 28], "uo": [1, 2, 8, 9], "place": [1, 29], "separ": [1, 22], "directori": [1, 2, 31], "within": [1, 2, 7, 22, 25, 28, 29], "exact": [1, 3, 8, 21, 22], "all": [1, 2, 7, 17, 24, 28], "program": [1, 17], "object": [1, 17, 22, 24, 29, 30], "function": [1, 17, 22, 31], "etc": [1, 7], "properli": 1, "us": [1, 2, 7, 8, 17, 24, 25, 28, 30], "system": [1, 31], "sphinx": [1, 2, 8], "restructuredtext": 1, "doc": [1, 2], "comment": 1, "code": [1, 2, 17, 28, 30], "have": [1, 17, 22], "web": 1, "page": [1, 8], "where": [1, 17, 22], "describ": [1, 22, 24, 25, 28, 29, 30], "There": [1, 7], "link": 1, "toward": 1, "explain": 1, "cover": 1, "unit": [1, 2, 8], "test": [1, 2, 8], "under": 1, "In": [1, 7], "order": 1, "do": 1, "so": [1, 22, 24], "unittest": [1, 2, 8], "exampl": [1, 2], "least": 1, "one": [1, 7, 22], "those": 1, "root": 1, "folder": [1, 2], "file": [1, 2, 9, 17, 21, 22, 29], "name": [1, 21, 22, 25, 29, 30], "py": 1, "vladimir": 1, "filipovi\u0107": 1, "vladofilipov": 1, "hotmail": 1, "com": 1, "variabl": [1, 3, 5, 8, 28], "neighborhood": [1, 3, 5, 8, 28], "search": [1, 3, 5, 8, 17, 28], "instal": 2, "provid": 2, "sourc": [2, 22], "poetri": [2, 8], "pip": 2, "check": [2, 22, 24], "successfulli": 2, "creat": [2, 24, 28], "virtual": 2, "environ": 2, "alreadi": [2, 17, 30], "same": [2, 7, 17, 22, 24, 25, 28, 29, 30], "command": [2, 17], "activ": 2, "deactiv": 2, "shell": 2, "project": [2, 8], "": [2, 17, 30], "packet": 2, "builder": 2, "depend": [2, 17], "execut": [2, 17, 22, 24, 25, 28], "develop": [2, 8], "applic": [2, 7], "strip": 2, "illustr": 2, "max": [2, 8, 17, 18], "Ones": [2, 8, 17, 18], "binari": [2, 17], "represent": [2, 17, 19, 21, 22, 24, 25, 28, 29, 30], "int": [2, 17, 19, 22, 24, 25, 28, 29, 30], "onli": [2, 7], "python": [2, 8], "example_vns_maxones_int_vns_exec": 2, "enhanc": 2, "both": [2, 7, 22], "app": 2, "example_vns_maxones_int_vns_exec_build": 2, "run": 2, "m": [2, 7], "build": 2, "apidoc": 2, "o": 2, "opt": [2, 8, 9], "chang": [2, 7], "current": [2, 7, 22, 24, 25, 28, 29, 30], "cd": 2, "clean": 2, "previous": 2, "html": 2, "make": 2, "total": [3, 4, 8, 24], "enumer": [3, 4, 8, 24], "basic": [3, 5, 8], "inform": [3, 5], "note": [3, 5, 8], "refer": [3, 5, 8], "denot": 7, "propos": 7, "mladenovi\u0107": 7, "hansen": 7, "mlahan1997": 7, "method": [7, 8], "solv": [7, 8, 17, 22, 24, 28, 30], "combinatori": [7, 10, 12], "global": [7, 10, 12], "It": 7, "explor": 7, "distant": 7, "incumb": 7, "solut": [7, 17, 19, 22, 24, 28, 30], "move": 7, "new": [7, 17, 22, 24, 25, 28, 29, 30], "an": 7, "improv": [7, 17, 28], "wa": 7, "made": 7, "The": [7, 17, 22, 24, 25, 28, 29, 30, 31], "local": [7, 17, 28], "appli": 7, "repeatedli": 7, "get": 7, "optima": 7, "systemat": 7, "two": [7, 17, 30], "phase": 7, "firstli": 7, "descent": 7, "find": 7, "optimum": 7, "final": 7, "perturb": 7, "out": 7, "correspond": 7, "vallei": 7, "ar": [7, 22, 30], "rapidli": 7, "increas": 7, "number": [7, 22, 24, 30], "pertain": 7, "mani": 7, "field": [7, 17, 21, 22, 24, 25, 28, 29, 30], "locat": 7, "theori": 7, "cluster": 7, "analysi": 7, "schedul": 7, "vehicl": 7, "rout": 7, "network": 7, "design": 7, "lot": 7, "size": [7, 30], "artifici": 7, "intellig": 7, "engin": 7, "pool": 7, "biologi": 7, "phylogeni": 7, "reliabl": 7, "geometri": 7, "telecommun": 7, "sever": [7, 28], "book": 7, "import": 7, "understand": 7, "like": 7, "handbook": 7, "genpot2010": 7, "glokoc2003": 7, "methodologi": 7, "burken2005": 7, "accord": 7, "autor": 7, "perform": 7, "procedur": [7, 17, 28], "minima": 7, "escap": 7, "which": [7, 17, 22, 30], "contain": [7, 8, 17, 28, 31], "them": 7, "built": 7, "upon": 7, "follow": [7, 22], "percept": 7, "A": 7, "minimum": 7, "respect": 7, "necessarili": 7, "anoth": 7, "possibl": 7, "For": 7, "rel": 7, "close": 7, "each": 7, "other": 7, "main": [7, 16, 17, 25, 28], "step": [7, 22], "through": 7, "choos": 7, "x": [7, 17, 24, 25, 28], "discov": 7, "direct": 7, "given": [7, 19], "k": [7, 17, 28], "proceed": 7, "f": 7, "If": 7, "heurist": 7, "stop": 7, "otherwis": 7, "iter": [7, 21, 22, 24, 25, 28], "usual": 7, "highest": 7, "also": 7, "relat": 7, "best": [7, 17, 22, 28], "shake": [7, 16, 17, 25, 28], "typic": 7, "involv": 7, "random": [7, 16, 17, 22, 24, 25, 28, 30], "extract": 7, "feasibl": [7, 17, 30], "compar": 7, "valu": [7, 17, 22, 24, 25, 28, 29, 30], "obtain": [7, 17, 22, 30], "return": [7, 17, 22, 24, 25, 28, 29, 30], "its": [7, 17, 30], "updat": [7, 22], "next": [7, 24], "consid": 7, "depth": 7, "variou": [7, 8, 17], "variant": [7, 17, 28], "vnsoptim": [7, 25, 28], "n": [7, 17], "p": 7, "1997": 7, "comput": 7, "oper": [7, 22, 24], "research": 7, "24": 7, "11": 7, "1097": 7, "1100": 7, "citeseerx": 7, "10": 7, "800": 7, "1797": 7, "doi": 7, "1016": 7, "s0305": 7, "0548": 7, "97": 7, "00031": 7, "gendreau": 7, "potvin": 7, "j": 7, "y": 7, "2010": 7, "springer": 7, "glover": 7, "kochenberg": 7, "g": [7, 30], "2003": 7, "kluwer": 7, "academ": 7, "publish": 7, "burk": 7, "ek": 7, "kendal": 7, "2005": 7, "edmund": 7, "graham": 7, "ed": 7, "introductori": 7, "tutori": 7, "decis": 7, "techniqu": [7, 17], "1007": 7, "978": 7, "4614": 7, "6940": 7, "7": 7, "isbn": 7, "6939": 7, "languag": 8, "tool": 8, "manag": 8, "gener": [8, 22, 24, 28, 30], "packag": [8, 9], "subpackag": [8, 9], "target_problem": [8, 9, 21, 22, 24, 25, 28, 30], "target_solut": [8, 9, 21, 24], "util": [8, 9, 21], "modul": [8, 9], "content": [8, 9], "multi_object": [8, 9, 10], "single_object": [8, 9, 10], "contributor": 8, "index": 8, "submodul": [9, 12, 16, 21], "output_control": [9, 21, 24, 25, 28], "evaluation_cache_control_statist": [9, 21], "logger": [9, 21], "trivial": [10, 12], "traveling_salesperson_problem": [12, 13], "max_ones_problem": [12, 16], "command_lin": [12, 16], "max_ones_problem_binary_bit_array_solut": [12, 16], "max_ones_problem_binary_bit_array_solution_vns_support": [12, 16], "max_ones_problem_binary_int_solut": [12, 16], "max_ones_problem_binary_int_solution_vns_support": [12, 16], "solver": [12, 16], "parse_argu": [16, 17], "maxonesproblem": [16, 17, 19], "copi": [16, 17, 21, 22, 24, 25, 28, 29, 30], "load_from_fil": [16, 17, 21, 29], "string_rep": [16, 17, 21, 22, 24, 25, 28, 29, 30], "maxonesproblembinarybitarraysolut": [16, 17, 19], "calculate_quality_directli": [16, 17, 21, 30], "copy_to": [16, 17, 21, 30], "init_from": [16, 17, 21, 30], "init_random": [16, 17, 21, 30], "native_represent": [16, 17, 21, 30], "representation_distance_directli": [16, 17, 21, 30], "string_represent": [16, 17, 21, 22, 30], "maxonesproblembinarybitarraysolutionvnssupport": [16, 17], "local_search_best_improv": [16, 17, 25, 28], "local_search_first_improv": [16, 17, 25, 28], "maxonesproblembinaryintsolut": [16, 17, 19], "maxonesproblembinaryintsolutionvnssupport": [16, 17], "paramet": [17, 22, 24, 25, 28, 29, 30], "ones": 17, "pars": 17, "repres": [17, 19, 24, 25, 28, 30], "file_path": [17, 21, 29], "str": [17, 22, 24, 25, 28, 29, 30, 31], "none": [17, 22, 24, 25, 28, 29, 30, 31], "dim": 17, "base": [17, 22, 24, 25, 28, 29, 30], "targetproblem": [17, 21, 22, 24, 25, 28, 29, 30], "instanc": [17, 22, 24, 25, 28, 29, 30], "properti": [17, 22, 24, 25, 28, 29, 30], "type": [17, 22, 24, 25, 28, 29, 30], "data_format": 17, "txt": 17, "read": [17, 29], "target": [17, 22, 29, 30], "data": [17, 22, 29], "format": 17, "delimit": [17, 22, 24, 25, 28, 29, 30], "indent": [17, 22, 24, 25, 28, 29, 30], "indentation_symbol": [17, 22, 24, 25, 28, 29, 30], "group_start": [17, 22, 24, 25, 28, 29, 30], "group_end": [17, 22, 24, 25, 28, 29, 30], "string": [17, 22, 24, 25, 28, 29, 30], "maxoneproblem": 17, "between": [17, 22, 24, 25, 28, 29, 30], "option": [17, 22, 24, 25, 28, 29, 30], "default": [17, 22, 24, 25, 28, 29, 30], "level": [17, 22, 24, 25, 28, 29, 30], "symbol": [17, 22, 24, 25, 28, 29, 30], "group": [17, 22, 24, 25, 28, 29, 30], "start": [17, 22, 24, 25, 28, 29, 30], "end": [17, 22, 24, 25, 28, 29, 30], "control": [17, 22, 24, 25, 28, 29, 30], "output": [17, 22, 24, 25, 28, 29, 30], "bitarrai": [17, 19], "ha": 17, "been": 17, "random_se": [17, 22, 25, 28, 30], "evaluation_cache_is_us": [17, 30], "bool": [17, 22, 24, 28, 29, 30], "fals": [17, 22, 30], "evaluation_cache_max_s": [17, 30], "distance_calculation_cache_is_us": [17, 30], "distance_calculation_cache_max_s": [17, 30], "targetsolut": [17, 21, 22, 24, 28, 30], "qualityofsolut": [17, 21, 30], "fit": [17, 22, 30], "calcul": [17, 25, 30], "nativ": [17, 30], "whose": [17, 31], "destin": [17, 30], "exist": [17, 30, 31], "ste": [17, 30], "representation_str": [17, 30], "solution_code_2": 17, "float": [17, 25, 30], "distanc": [17, 30], "determin": [17, 22, 30], "solution_code_1": 17, "first": [17, 22, 28, 30], "second": [17, 22, 25, 30], "interv": [17, 24, 25, 28], "part": [17, 28], "problemsolutionvnssupport": [17, 25, 28], "result": [17, 28], "doe": [17, 28], "differ": [17, 28], "more": [17, 28], "than": [17, 22, 28], "insid": [17, 28], "shakingpoint": 17, "success": [17, 28], "integ": 17, "directli": [17, 30], "representation_1": [17, 30], "r_co": [17, 24, 28, 30], "representation_2": [17, 30], "thi": [17, 22, 24, 25], "line": 17, "solution_code_distance_cache_control_statist": [21, 22], "best_solut": [21, 22], "copy_to_best_solut": [21, 22], "evalu": [21, 22, 24, 30], "execution_end": [21, 22], "execution_start": [21, 22], "init": [21, 22, 24, 25, 28], "is_first_solution_bett": [21, 22], "write_output_headers_if_need": [21, 22], "write_output_values_if_need": [21, 22], "outputcontrol": [21, 22, 24, 25, 28], "fields_definit": [21, 22], "fields_head": [21, 22], "moment": [21, 22], "output_fil": [21, 22], "write_after_algorithm": [21, 22], "write_after_evalu": [21, 22], "write_after_iter": [21, 22], "write_after_step_in_iter": [21, 22], "write_before_algorithm": [21, 22], "write_before_evalu": [21, 22], "write_before_iter": [21, 22], "write_before_step_in_iter": [21, 22], "write_to_output": [21, 22], "dimens": [21, 29], "is_minim": [21, 29], "evaluationcachecontrolstatist": [21, 30], "cach": [21, 30], "cache_hit_count": [21, 30], "cache_request_count": [21, 30], "increment_cache_hit_count": [21, 30], "increment_cache_request_count": [21, 30], "is_cach": [21, 30], "max_cache_s": [21, 30], "fitness_valu": [21, 22, 30], "is_feas": [21, 22, 30], "objective_valu": [21, 22, 30], "calculate_qu": [21, 30], "representation_dist": [21, 30], "ensure_dir": [21, 31], "electro_magnetism_like_metaheusrist": [22, 25], "genetic_algorithm": [22, 25], "variable_neighborhood_search": [22, 25], "problem_solution_vns_support": [22, 25], "vns_optim": [22, 25], "additional_statistics_control": [22, 25, 28], "elapsed_second": [22, 25], "finish_control": [22, 25, 28], "main_loop": [22, 25], "main_loop_iter": [22, 25, 28], "getter": [22, 24, 25, 28, 29, 30], "dure": [22, 24, 25, 30], "far": 22, "abstract": [22, 24, 25, 28, 29, 30], "argument": 22, "becom": 22, "info": 22, "about": 22, "time": [22, 25], "when": 22, "cope": 22, "datetim": 22, "sol1": 22, "sol2": 22, "better": 22, "true": 22, "wors": 22, "equal": 22, "write": 22, "header": 22, "necessari": 22, "step_nam": 22, "step_name_valu": 22, "written": 22, "after_algorithm": 22, "before_algorithm": 22, "after_iter": 22, "before_iter": 22, "after_evalu": 22, "before_evalu": 22, "after_step_in_iter": 22, "before_step_in_iter": 22, "what": 22, "instead": 22, "textiowrapp": 22, "comma": 22, "list": [22, 28, 30], "definit": 22, "head": 22, "after": 22, "prior": 22, "problemsolutiontesupport": 24, "can_progress": 24, "process": 24, "indic": 24, "overall_number_of_evalu": 24, "overal": [24, 30], "requir": 24, "finish": [24, 25], "progress": 24, "intern": 24, "counter": 24, "configur": 24, "taken": 24, "consider": 24, "state": 24, "reflect": 24, "reset": 24, "over": 24, "totalenumer": 24, "teoptim": 24, "initial_solut": [24, 28], "current_solut": 24, "subtyp": 24, "classmethod": [24, 28], "from_construction_tupl": [24, 25, 28], "construction_tupl": [24, 28], "teoptimizerconstructionparamet": 24, "addit": [24, 28], "constructor": [24, 28], "tupl": [24, 28, 30], "te": 24, "totalenumerationconstructorparamet": 24, "k_max": [25, 28], "k_min": [25, 28], "finishcontrol": [25, 28], "additionalstatisticscontrol": [25, 28], "keep": 25, "statist": [25, 30], "elaps": 25, "criteria": 25, "loop": [25, 28], "One": [25, 28], "seed": 25, "solution_represent": 28, "suppli": 28, "collect": 28, "shaken": 28, "local_search_typ": 28, "singlesolutionmetaheurist": 28, "encapsul": 28, "vnsoptimizerconstructionparamet": 28, "path": [29, 31], "minim": 29, "take": 29, "intrins": 29, "data_represent": 29, "dict": 30, "hit": 30, "increment": 30, "maximum": 30, "unlimit": 30, "alia": 30, "consult": 30, "e": 30, "deal": 31, "path_to_dir": 31, "ensur": 31, "specif": 31, "param": 31, "log": 31}, "objects": {"": [[10, 0, 0, "-", "opt"], [21, 0, 0, "-", "uo"]], "opt": [[11, 0, 0, "-", "multi_objective"], [12, 0, 0, "-", "single_objective"]], "opt.single_objective": [[13, 0, 0, "-", "combinatorial"], [15, 0, 0, "-", "global"], [16, 0, 0, "-", "trivial"]], "opt.single_objective.combinatorial": [[14, 0, 0, "-", "traveling_salesperson_problem"]], "opt.single_objective.trivial": [[17, 0, 0, "-", "max_ones_problem"]], "opt.single_objective.trivial.max_ones_problem": [[17, 0, 0, "-", "command_line"], [17, 0, 0, "-", "max_ones_problem"], [17, 0, 0, "-", "max_ones_problem_binary_bit_array_solution"], [17, 0, 0, "-", "max_ones_problem_binary_bit_array_solution_vns_support"], [17, 0, 0, "-", "max_ones_problem_binary_int_solution"], [17, 0, 0, "-", "max_ones_problem_binary_int_solution_vns_support"], [17, 0, 0, "-", "solver"]], "opt.single_objective.trivial.max_ones_problem.command_line": [[17, 1, 1, "", "parse_arguments"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem": [[17, 2, 1, "", "MaxOnesProblem"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem.MaxOnesProblem": [[17, 3, 1, "", "copy"], [17, 3, 1, "", "load_from_file"], [17, 3, 1, "", "string_rep"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution": [[17, 2, 1, "", "MaxOnesProblemBinaryBitArraySolution"], [17, 1, 1, "", "random"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution": [[17, 3, 1, "", "calculate_quality_directly"], [17, 3, 1, "", "copy"], [17, 3, 1, "", "copy_to"], [17, 3, 1, "", "init_from"], [17, 3, 1, "", "init_random"], [17, 3, 1, "", "native_representation"], [17, 3, 1, "", "representation_distance_directly"], [17, 3, 1, "", "string_rep"], [17, 3, 1, "", "string_representation"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support": [[17, 2, 1, "", "MaxOnesProblemBinaryBitArraySolutionVnsSupport"], [17, 1, 1, "", "random"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport": [[17, 3, 1, "", "copy"], [17, 3, 1, "", "local_search_best_improvement"], [17, 3, 1, "", "local_search_first_improvement"], [17, 3, 1, "", "shaking"], [17, 3, 1, "", "string_rep"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution": [[17, 2, 1, "", "MaxOnesProblemBinaryIntSolution"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution": [[17, 3, 1, "", "calculate_quality_directly"], [17, 3, 1, "", "copy"], [17, 3, 1, "", "copy_to"], [17, 3, 1, "", "init_from"], [17, 3, 1, "", "init_random"], [17, 3, 1, "", "native_representation"], [17, 3, 1, "", "representation_distance_directly"], [17, 3, 1, "", "string_rep"], [17, 3, 1, "", "string_representation"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support": [[17, 2, 1, "", "MaxOnesProblemBinaryIntSolutionVnsSupport"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport": [[17, 3, 1, "", "copy"], [17, 3, 1, "", "local_search_best_improvement"], [17, 3, 1, "", "local_search_first_improvement"], [17, 3, 1, "", "shaking"], [17, 3, 1, "", "string_rep"]], "opt.single_objective.trivial.max_ones_problem.solver": [[17, 1, 1, "", "main"]], "uo": [[22, 0, 0, "-", "algorithm"], [29, 0, 0, "-", "target_problem"], [30, 0, 0, "-", "target_solution"], [31, 0, 0, "-", "utils"]], "uo.algorithm": [[22, 0, 0, "-", "algorithm"], [23, 0, 0, "-", "exact"], [25, 0, 0, "-", "metaheuristic"], [22, 0, 0, "-", "output_control"]], "uo.algorithm.algorithm": [[22, 2, 1, "", "Algorithm"]], "uo.algorithm.algorithm.Algorithm": [[22, 4, 1, "", "best_solution"], [22, 3, 1, "", "copy"], [22, 3, 1, "", "copy_to_best_solution"], [22, 4, 1, "", "evaluation"], [22, 4, 1, "", "execution_ended"], [22, 4, 1, "", "execution_started"], [22, 3, 1, "", "init"], [22, 3, 1, "", "is_first_solution_better"], [22, 4, 1, "", "iteration"], [22, 4, 1, "", "name"], [22, 4, 1, "", "output_control"], [22, 3, 1, "", "string_rep"], [22, 4, 1, "", "target_problem"], [22, 3, 1, "", "write_output_headers_if_needed"], [22, 3, 1, "", "write_output_values_if_needed"]], "uo.algorithm.exact": [[24, 0, 0, "-", "total_enumeration"]], "uo.algorithm.exact.total_enumeration": [[24, 0, 0, "-", "problem_solution_te_support"], [24, 0, 0, "-", "te_optimizer"], [24, 0, 0, "-", "te_optimizer_constructor_parameters"]], "uo.algorithm.exact.total_enumeration.problem_solution_te_support": [[24, 2, 1, "", "ProblemSolutionTeSupport"]], "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport": [[24, 3, 1, "", "can_progress"], [24, 3, 1, "", "overall_number_of_evaluations"], [24, 3, 1, "", "progress"], [24, 3, 1, "", "reset"]], "uo.algorithm.exact.total_enumeration.te_optimizer": [[24, 2, 1, "", "TeOptimizer"], [24, 1, 1, "", "random"]], "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer": [[24, 3, 1, "", "copy"], [24, 4, 1, "", "current_solution"], [24, 3, 1, "", "from_construction_tuple"], [24, 3, 1, "", "init"], [24, 4, 1, "", "iteration"], [24, 3, 1, "", "optimize"], [24, 3, 1, "", "string_rep"]], "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters": [[24, 2, 1, "", "TeOptimizerConstructionParameters"]], "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters": [[24, 5, 1, "", "initial_solution"], [24, 5, 1, "", "output_control"], [24, 5, 1, "", "problem_solution_te_support"], [24, 5, 1, "", "target_problem"]], "uo.algorithm.metaheuristic": [[26, 0, 0, "-", "electro_magnetism_like_metaheusristic"], [27, 0, 0, "-", "genetic_algorithm"], [25, 0, 0, "-", "metaheuristic"], [28, 0, 0, "-", "variable_neighborhood_search"]], "uo.algorithm.metaheuristic.metaheuristic": [[25, 2, 1, "", "Metaheuristic"], [25, 1, 1, "", "random"]], "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic": [[25, 4, 1, "", "additional_statistics_control"], [25, 3, 1, "", "copy"], [25, 3, 1, "", "elapsed_seconds"], [25, 4, 1, "", "finish_control"], [25, 3, 1, "", "main_loop"], [25, 3, 1, "", "main_loop_iteration"], [25, 3, 1, "", "optimize"], [25, 4, 1, "", "random_seed"], [25, 3, 1, "", "string_rep"]], "uo.algorithm.metaheuristic.variable_neighborhood_search": [[28, 0, 0, "-", "problem_solution_vns_support"], [28, 0, 0, "-", "vns_optimizer"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support": [[28, 2, 1, "", "ProblemSolutionVnsSupport"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport": [[28, 3, 1, "", "local_search_best_improvement"], [28, 3, 1, "", "local_search_first_improvement"], [28, 3, 1, "", "shaking"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer": [[28, 2, 1, "", "VnsOptimizer"], [28, 1, 1, "", "random"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer": [[28, 3, 1, "", "copy"], [28, 3, 1, "", "from_construction_tuple"], [28, 3, 1, "", "init"], [28, 4, 1, "", "k_max"], [28, 4, 1, "", "k_min"], [28, 3, 1, "", "main_loop_iteration"], [28, 3, 1, "", "string_rep"]], "uo.algorithm.output_control": [[22, 2, 1, "", "OutputControl"]], "uo.algorithm.output_control.OutputControl": [[22, 4, 1, "", "fields"], [22, 4, 1, "", "fields_definitions"], [22, 4, 1, "", "fields_headings"], [22, 4, 1, "", "moments"], [22, 4, 1, "", "output_file"], [22, 3, 1, "", "string_rep"], [22, 4, 1, "", "write_after_algorithm"], [22, 4, 1, "", "write_after_evaluation"], [22, 4, 1, "", "write_after_iteration"], [22, 4, 1, "", "write_after_step_in_iteration"], [22, 4, 1, "", "write_before_algorithm"], [22, 4, 1, "", "write_before_evaluation"], [22, 4, 1, "", "write_before_iteration"], [22, 4, 1, "", "write_before_step_in_iteration"], [22, 4, 1, "", "write_to_output"]], "uo.target_problem": [[29, 0, 0, "-", "target_problem"]], "uo.target_problem.target_problem": [[29, 2, 1, "", "TargetProblem"]], "uo.target_problem.target_problem.TargetProblem": [[29, 3, 1, "", "copy"], [29, 4, 1, "", "dimension"], [29, 4, 1, "", "file_path"], [29, 4, 1, "", "is_minimization"], [29, 3, 1, "", "load_from_file"], [29, 4, 1, "", "name"], [29, 3, 1, "", "string_rep"]], "uo.target_solution": [[30, 0, 0, "-", "evaluation_cache_control_statistics"], [30, 0, 0, "-", "target_solution"]], "uo.target_solution.evaluation_cache_control_statistics": [[30, 2, 1, "", "EvaluationCacheControlStatistics"]], "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics": [[30, 4, 1, "", "cache"], [30, 4, 1, "", "cache_hit_count"], [30, 4, 1, "", "cache_request_count"], [30, 3, 1, "", "copy"], [30, 3, 1, "", "increment_cache_hit_count"], [30, 3, 1, "", "increment_cache_request_count"], [30, 4, 1, "", "is_caching"], [30, 4, 1, "", "max_cache_size"], [30, 3, 1, "", "string_rep"]], "uo.target_solution.target_solution": [[30, 2, 1, "", "QualityOfSolution"], [30, 2, 1, "", "TargetSolution"]], "uo.target_solution.target_solution.QualityOfSolution": [[30, 5, 1, "", "fitness_value"], [30, 5, 1, "", "is_feasible"], [30, 5, 1, "", "objective_value"]], "uo.target_solution.target_solution.TargetSolution": [[30, 3, 1, "", "calculate_quality"], [30, 3, 1, "", "calculate_quality_directly"], [30, 3, 1, "", "copy"], [30, 3, 1, "", "copy_to"], [30, 3, 1, "", "evaluate"], [30, 4, 1, "", "fitness_value"], [30, 3, 1, "", "init_from"], [30, 3, 1, "", "init_random"], [30, 4, 1, "", "is_feasible"], [30, 4, 1, "", "name"], [30, 3, 1, "", "native_representation"], [30, 4, 1, "", "objective_value"], [30, 4, 1, "", "representation"], [30, 3, 1, "", "representation_distance"], [30, 3, 1, "", "representation_distance_directly"], [30, 3, 1, "", "string_rep"], [30, 3, 1, "", "string_representation"]], "uo.utils": [[31, 0, 0, "-", "files"], [31, 0, 0, "-", "logger"]], "uo.utils.files": [[31, 1, 1, "", "ensure_dir"]]}, "objtypes": {"0": "py:module", "1": "py:function", "2": "py:class", "3": "py:method", "4": "py:property", "5": "py:attribute"}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "function", "Python function"], "2": ["py", "class", "Python class"], "3": ["py", "method", "Python method"], "4": ["py", "property", "Python property"], "5": ["py", "attribute", "Python attribute"]}, "titleterms": {"changelog": [0, 8], "how": [1, 2, 8], "contribut": [1, 8], "contributor": 1, "guid": [2, 8], "implement": [3, 4, 5, 7], "algorithm": [3, 4, 5, 7, 8, 22, 23, 24, 25, 26, 27, 28], "exact": [4, 23, 24], "metaheurist": [5, 25, 26, 27, 28], "total": 6, "enumer": 6, "variabl": 7, "neighborhood": 7, "search": 7, "basic": 7, "inform": [7, 8], "structur": 7, "note": 7, "refer": 7, "univers": 8, "optim": [8, 9], "librari": [8, 9], "tutori": [8, 20], "To": 8, "In": 8, "depth": 8, "program": 8, "code": 8, "indic": 8, "tabl": 8, "opt": [10, 11, 12, 13, 14, 15, 16, 17], "packag": [10, 11, 12, 13, 14, 15, 16, 17, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31], "subpackag": [10, 12, 13, 16, 21, 22, 25], "modul": [10, 11, 12, 13, 14, 15, 16, 17, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31], "content": [10, 11, 12, 13, 14, 15, 16, 17, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31], "multi_object": 11, "single_object": [12, 13, 14, 15, 16, 17], "combinatori": [13, 14], "traveling_salesperson_problem": 14, "global": 15, "trivial": [16, 17], "max_ones_problem": 17, "submodul": [17, 22, 24, 25, 28, 29, 30, 31], "command_lin": 17, "max_ones_problem_binary_bit_array_solut": 17, "max_ones_problem_binary_bit_array_solution_vns_support": 17, "max_ones_problem_binary_int_solut": 17, "max_ones_problem_binary_int_solution_vns_support": 17, "solver": 17, "problem": [18, 19], "solv": 18, "max": 19, "Ones": 19, "uo": [21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31], "output_control": 22, "total_enumer": 24, "problem_solution_te_support": 24, "te_optim": 24, "te_optimizer_constructor_paramet": 24, "solution_code_distance_cache_control_statist": 25, "electro_magnetism_like_metaheusrist": 26, "genetic_algorithm": 27, "variable_neighborhood_search": 28, "problem_solution_vns_support": 28, "vns_optim": 28, "target_problem": 29, "target_solut": 30, "evaluation_cache_control_statist": 30, "util": 31, "file": 31, "logger": 31}, "envversion": {"sphinx.domains.c": 3, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 9, "sphinx.domains.index": 1, "sphinx.domains.javascript": 3, "sphinx.domains.math": 2, "sphinx.domains.python": 4, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.intersphinx": 1, "sphinx": 60}, "alltitles": {"Changelog": [[0, "changelog"]], "How to Contribute": [[1, "how-to-contribute"]], "Contributors": [[1, "contributors"]], "How-to Guides": [[2, "how-to-guides"]], "Implemented algorithms": [[3, "implemented-algorithms"], [3, null]], "Implemented exact algorithms": [[4, "implemented-exact-algorithms"], [4, null]], "Implemented metaheuristic algorithms": [[5, "implemented-metaheuristic-algorithms"], [5, null]], "Total Enumeration": [[6, "total-enumeration"]], "Variable Neighborhood Search": [[7, "variable-neighborhood-search"]], "Basic information": [[7, "basic-information"]], "Structure of the algorithm": [[7, "structure-of-the-algorithm"]], "Implementation notes": [[7, "implementation-notes"]], "References": [[7, "references"]], "Universal optimizer - Optimization algorithms library": [[8, "universal-optimizer-optimization-algorithms-library"]], "Tutorials and How-To Guides": [[8, null]], "In-depth Information": [[8, null]], "Program code": [[8, null]], "Changelog and Contribution": [[8, null]], "Indices and tables": [[8, "indices-and-tables"]], "Library": [[9, "library"]], "Optimization": [[9, "optimization"]], "opt package": [[10, "opt-package"]], "Subpackages": [[10, "subpackages"], [12, "subpackages"], [13, "subpackages"], [16, "subpackages"], [21, "subpackages"], [22, "subpackages"], [25, "subpackages"]], "Module contents": [[10, "module-opt"], [11, "module-opt.multi_objective"], [12, "module-opt.single_objective"], [13, "module-opt.single_objective.combinatorial"], [14, "module-opt.single_objective.combinatorial.traveling_salesperson_problem"], [15, "module-opt.single_objective.global"], [16, "module-opt.single_objective.trivial"], [17, "module-opt.single_objective.trivial.max_ones_problem"], [21, "module-uo"], [22, "module-uo.algorithm"], [23, "module-uo.algorithm.exact"], [24, "module-uo.algorithm.exact.total_enumeration"], [25, "module-uo.algorithm.metaheuristic"], [26, "module-uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic"], [27, "module-uo.algorithm.metaheuristic.genetic_algorithm"], [28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search"], [29, "module-uo.target_problem"], [30, "module-uo.target_solution"], [31, "module-uo.utils"]], "opt.multi_objective package": [[11, "opt-multi-objective-package"]], "opt.single_objective package": [[12, "opt-single-objective-package"]], "opt.single_objective.combinatorial package": [[13, "opt-single-objective-combinatorial-package"]], "opt.single_objective.combinatorial.traveling_salesperson_problem package": [[14, "opt-single-objective-combinatorial-traveling-salesperson-problem-package"]], "opt.single_objective.global package": [[15, "opt-single-objective-global-package"]], "opt.single_objective.trivial package": [[16, "opt-single-objective-trivial-package"]], "opt.single_objective.trivial.max_ones_problem package": [[17, "opt-single-objective-trivial-max-ones-problem-package"]], "Submodules": [[17, "submodules"], [22, "submodules"], [24, "submodules"], [25, "submodules"], [28, "submodules"], [29, "submodules"], [30, "submodules"], [31, "submodules"]], "opt.single_objective.trivial.max_ones_problem.command_line module": [[17, "module-opt.single_objective.trivial.max_ones_problem.command_line"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem module": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution module": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support module": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution module": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support module": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support"]], "opt.single_objective.trivial.max_ones_problem.solver module": [[17, "module-opt.single_objective.trivial.max_ones_problem.solver"]], "Problems to be solved": [[18, "problems-to-be-solved"]], "Problems": [[18, null]], "Max Ones Problem": [[19, "max-ones-problem"]], "Tutorials": [[20, "tutorials"]], "uo package": [[21, "uo-package"]], "uo.algorithm package": [[22, "uo-algorithm-package"]], "uo.algorithm.algorithm module": [[22, "module-uo.algorithm.algorithm"]], "uo.algorithm.output_control module": [[22, "module-uo.algorithm.output_control"]], "uo.algorithm.exact package": [[23, "uo-algorithm-exact-package"]], "uo.algorithm.exact.total_enumeration package": [[24, "uo-algorithm-exact-total-enumeration-package"]], "uo.algorithm.exact.total_enumeration.problem_solution_te_support module": [[24, "module-uo.algorithm.exact.total_enumeration.problem_solution_te_support"]], "uo.algorithm.exact.total_enumeration.te_optimizer module": [[24, "module-uo.algorithm.exact.total_enumeration.te_optimizer"]], "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters module": [[24, "module-uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters"]], "uo.algorithm.metaheuristic package": [[25, "uo-algorithm-metaheuristic-package"]], "uo.algorithm.metaheuristic.metaheuristic module": [[25, "module-uo.algorithm.metaheuristic.metaheuristic"]], "uo.algorithm.metaheuristic.solution_code_distance_cache_control_statistics module": [[25, "uo-algorithm-metaheuristic-solution-code-distance-cache-control-statistics-module"]], "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic package": [[26, "uo-algorithm-metaheuristic-electro-magnetism-like-metaheusristic-package"]], "uo.algorithm.metaheuristic.genetic_algorithm package": [[27, "uo-algorithm-metaheuristic-genetic-algorithm-package"]], "uo.algorithm.metaheuristic.variable_neighborhood_search package": [[28, "uo-algorithm-metaheuristic-variable-neighborhood-search-package"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support module": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer module": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer"]], "uo.target_problem package": [[29, "uo-target-problem-package"]], "uo.target_problem.target_problem module": [[29, "module-uo.target_problem.target_problem"]], "uo.target_solution package": [[30, "uo-target-solution-package"]], "uo.target_solution.evaluation_cache_control_statistics module": [[30, "module-uo.target_solution.evaluation_cache_control_statistics"]], "uo.target_solution.target_solution module": [[30, "module-uo.target_solution.target_solution"]], "uo.utils package": [[31, "uo-utils-package"]], "uo.utils.files module": [[31, "module-uo.utils.files"]], "uo.utils.logger module": [[31, "module-uo.utils.logger"]]}, "indexentries": {"module": [[10, "module-opt"], [11, "module-opt.multi_objective"], [12, "module-opt.single_objective"], [13, "module-opt.single_objective.combinatorial"], [14, "module-opt.single_objective.combinatorial.traveling_salesperson_problem"], [15, "module-opt.single_objective.global"], [16, "module-opt.single_objective.trivial"], [17, "module-opt.single_objective.trivial.max_ones_problem"], [17, "module-opt.single_objective.trivial.max_ones_problem.command_line"], [17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem"], [17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution"], [17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support"], [17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution"], [17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support"], [17, "module-opt.single_objective.trivial.max_ones_problem.solver"], [21, "module-uo"], [22, "module-uo.algorithm"], [22, "module-uo.algorithm.algorithm"], [22, "module-uo.algorithm.output_control"], [23, "module-uo.algorithm.exact"], [24, "module-uo.algorithm.exact.total_enumeration"], [24, "module-uo.algorithm.exact.total_enumeration.problem_solution_te_support"], [24, "module-uo.algorithm.exact.total_enumeration.te_optimizer"], [24, "module-uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters"], [25, "module-uo.algorithm.metaheuristic"], [25, "module-uo.algorithm.metaheuristic.metaheuristic"], [26, "module-uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic"], [27, "module-uo.algorithm.metaheuristic.genetic_algorithm"], [28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search"], [28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support"], [28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer"], [29, "module-uo.target_problem"], [29, "module-uo.target_problem.target_problem"], [30, "module-uo.target_solution"], [30, "module-uo.target_solution.evaluation_cache_control_statistics"], [30, "module-uo.target_solution.target_solution"], [31, "module-uo.utils"], [31, "module-uo.utils.files"], [31, "module-uo.utils.logger"]], "opt": [[10, "module-opt"]], "opt.multi_objective": [[11, "module-opt.multi_objective"]], "opt.single_objective": [[12, "module-opt.single_objective"]], "opt.single_objective.combinatorial": [[13, "module-opt.single_objective.combinatorial"]], "opt.single_objective.combinatorial.traveling_salesperson_problem": [[14, "module-opt.single_objective.combinatorial.traveling_salesperson_problem"]], "opt.single_objective.global": [[15, "module-opt.single_objective.global"]], "opt.single_objective.trivial": [[16, "module-opt.single_objective.trivial"]], "maxonesproblem (class in opt.single_objective.trivial.max_ones_problem.max_ones_problem)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem.MaxOnesProblem"]], "maxonesproblembinarybitarraysolution (class in opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution"]], "maxonesproblembinarybitarraysolutionvnssupport (class in opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport"]], "maxonesproblembinaryintsolution (class in opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution"]], "maxonesproblembinaryintsolutionvnssupport (class in opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport"]], "calculate_quality_directly() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.calculate_quality_directly"]], "calculate_quality_directly() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.calculate_quality_directly"]], "copy() (opt.single_objective.trivial.max_ones_problem.max_ones_problem.maxonesproblem method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem.MaxOnesProblem.copy"]], "copy() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.copy"]], "copy() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.copy"]], "copy() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.copy"]], "copy() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.copy"]], "copy_to() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.copy_to"]], "copy_to() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.copy_to"]], "init_from() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.init_from"]], "init_from() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.init_from"]], "init_random() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.init_random"]], "init_random() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.init_random"]], "load_from_file() (opt.single_objective.trivial.max_ones_problem.max_ones_problem.maxonesproblem method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem.MaxOnesProblem.load_from_file"]], "local_search_best_improvement() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.local_search_best_improvement"]], "local_search_best_improvement() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.local_search_best_improvement"]], "local_search_first_improvement() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.local_search_first_improvement"]], "local_search_first_improvement() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.local_search_first_improvement"]], "main() (in module opt.single_objective.trivial.max_ones_problem.solver)": [[17, "opt.single_objective.trivial.max_ones_problem.solver.main"]], "native_representation() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.native_representation"]], "native_representation() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.native_representation"]], "opt.single_objective.trivial.max_ones_problem": [[17, "module-opt.single_objective.trivial.max_ones_problem"]], "opt.single_objective.trivial.max_ones_problem.command_line": [[17, "module-opt.single_objective.trivial.max_ones_problem.command_line"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution"]], "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support": [[17, "module-opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support"]], "opt.single_objective.trivial.max_ones_problem.solver": [[17, "module-opt.single_objective.trivial.max_ones_problem.solver"]], "parse_arguments() (in module opt.single_objective.trivial.max_ones_problem.command_line)": [[17, "opt.single_objective.trivial.max_ones_problem.command_line.parse_arguments"]], "random() (in module opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.random"]], "random() (in module opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.random"]], "representation_distance_directly() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.representation_distance_directly"]], "representation_distance_directly() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.representation_distance_directly"]], "shaking() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.shaking"]], "shaking() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.shaking"]], "string_rep() (opt.single_objective.trivial.max_ones_problem.max_ones_problem.maxonesproblem method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem.MaxOnesProblem.string_rep"]], "string_rep() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.string_rep"]], "string_rep() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.string_rep"]], "string_rep() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.string_rep"]], "string_rep() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.string_rep"]], "string_representation() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.string_representation"]], "string_representation() (opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.string_representation"]], "uo": [[21, "module-uo"]], "algorithm (class in uo.algorithm.algorithm)": [[22, "uo.algorithm.algorithm.Algorithm"]], "outputcontrol (class in uo.algorithm.output_control)": [[22, "uo.algorithm.output_control.OutputControl"]], "best_solution (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.best_solution"]], "copy() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.copy"]], "copy_to_best_solution() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.copy_to_best_solution"]], "evaluation (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.evaluation"]], "execution_ended (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.execution_ended"]], "execution_started (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.execution_started"]], "fields (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.fields"]], "fields_definitions (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.fields_definitions"]], "fields_headings (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.fields_headings"]], "init() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.init"]], "is_first_solution_better() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.is_first_solution_better"]], "iteration (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.iteration"]], "moments (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.moments"]], "name (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.name"]], "output_control (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.output_control"]], "output_file (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.output_file"]], "string_rep() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.string_rep"]], "string_rep() (uo.algorithm.output_control.outputcontrol method)": [[22, "uo.algorithm.output_control.OutputControl.string_rep"]], "target_problem (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.target_problem"]], "uo.algorithm": [[22, "module-uo.algorithm"]], "uo.algorithm.algorithm": [[22, "module-uo.algorithm.algorithm"]], "uo.algorithm.output_control": [[22, "module-uo.algorithm.output_control"]], "write_after_algorithm (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_after_algorithm"]], "write_after_evaluation (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_after_evaluation"]], "write_after_iteration (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_after_iteration"]], "write_after_step_in_iteration (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_after_step_in_iteration"]], "write_before_algorithm (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_before_algorithm"]], "write_before_evaluation (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_before_evaluation"]], "write_before_iteration (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_before_iteration"]], "write_before_step_in_iteration (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_before_step_in_iteration"]], "write_output_headers_if_needed() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.write_output_headers_if_needed"]], "write_output_values_if_needed() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.write_output_values_if_needed"]], "write_to_output (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_to_output"]], "uo.algorithm.exact": [[23, "module-uo.algorithm.exact"]], "problemsolutiontesupport (class in uo.algorithm.exact.total_enumeration.problem_solution_te_support)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport"]], "teoptimizer (class in uo.algorithm.exact.total_enumeration.te_optimizer)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer"]], "teoptimizerconstructionparameters (class in uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters"]], "can_progress() (uo.algorithm.exact.total_enumeration.problem_solution_te_support.problemsolutiontesupport method)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport.can_progress"]], "copy() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.copy"]], "current_solution (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer property)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.current_solution"]], "from_construction_tuple() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer class method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.from_construction_tuple"]], "init() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.init"]], "initial_solution (uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.teoptimizerconstructionparameters attribute)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters.initial_solution"]], "iteration (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer property)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.iteration"]], "optimize() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.optimize"]], "output_control (uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.teoptimizerconstructionparameters attribute)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters.output_control"]], "overall_number_of_evaluations() (uo.algorithm.exact.total_enumeration.problem_solution_te_support.problemsolutiontesupport method)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport.overall_number_of_evaluations"]], "problem_solution_te_support (uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.teoptimizerconstructionparameters attribute)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters.problem_solution_te_support"]], "progress() (uo.algorithm.exact.total_enumeration.problem_solution_te_support.problemsolutiontesupport method)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport.progress"]], "random() (in module uo.algorithm.exact.total_enumeration.te_optimizer)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.random"]], "reset() (uo.algorithm.exact.total_enumeration.problem_solution_te_support.problemsolutiontesupport method)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport.reset"]], "string_rep() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.string_rep"]], "target_problem (uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.teoptimizerconstructionparameters attribute)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters.target_problem"]], "uo.algorithm.exact.total_enumeration": [[24, "module-uo.algorithm.exact.total_enumeration"]], "uo.algorithm.exact.total_enumeration.problem_solution_te_support": [[24, "module-uo.algorithm.exact.total_enumeration.problem_solution_te_support"]], "uo.algorithm.exact.total_enumeration.te_optimizer": [[24, "module-uo.algorithm.exact.total_enumeration.te_optimizer"]], "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters": [[24, "module-uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters"]], "metaheuristic (class in uo.algorithm.metaheuristic.metaheuristic)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic"]], "additional_statistics_control (uo.algorithm.metaheuristic.metaheuristic.metaheuristic property)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.additional_statistics_control"]], "copy() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.copy"]], "elapsed_seconds() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.elapsed_seconds"]], "finish_control (uo.algorithm.metaheuristic.metaheuristic.metaheuristic property)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.finish_control"]], "main_loop() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.main_loop"]], "main_loop_iteration() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.main_loop_iteration"]], "optimize() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.optimize"]], "random() (in module uo.algorithm.metaheuristic.metaheuristic)": [[25, "uo.algorithm.metaheuristic.metaheuristic.random"]], "random_seed (uo.algorithm.metaheuristic.metaheuristic.metaheuristic property)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.random_seed"]], "string_rep() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.string_rep"]], "uo.algorithm.metaheuristic": [[25, "module-uo.algorithm.metaheuristic"]], "uo.algorithm.metaheuristic.metaheuristic": [[25, "module-uo.algorithm.metaheuristic.metaheuristic"]], "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic": [[26, "module-uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic"]], "uo.algorithm.metaheuristic.genetic_algorithm": [[27, "module-uo.algorithm.metaheuristic.genetic_algorithm"]], "problemsolutionvnssupport (class in uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport"]], "vnsoptimizer (class in uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer"]], "copy() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.copy"]], "from_construction_tuple() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer class method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.from_construction_tuple"]], "init() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.init"]], "k_max (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer property)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.k_max"]], "k_min (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer property)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.k_min"]], "local_search_best_improvement() (uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.problemsolutionvnssupport method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport.local_search_best_improvement"]], "local_search_first_improvement() (uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.problemsolutionvnssupport method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport.local_search_first_improvement"]], "main_loop_iteration() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.main_loop_iteration"]], "random() (in module uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.random"]], "shaking() (uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.problemsolutionvnssupport method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport.shaking"]], "string_rep() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.string_rep"]], "uo.algorithm.metaheuristic.variable_neighborhood_search": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer"]], "targetproblem (class in uo.target_problem.target_problem)": [[29, "uo.target_problem.target_problem.TargetProblem"]], "copy() (uo.target_problem.target_problem.targetproblem method)": [[29, "uo.target_problem.target_problem.TargetProblem.copy"]], "dimension (uo.target_problem.target_problem.targetproblem property)": [[29, "uo.target_problem.target_problem.TargetProblem.dimension"]], "file_path (uo.target_problem.target_problem.targetproblem property)": [[29, "uo.target_problem.target_problem.TargetProblem.file_path"]], "is_minimization (uo.target_problem.target_problem.targetproblem property)": [[29, "uo.target_problem.target_problem.TargetProblem.is_minimization"]], "load_from_file() (uo.target_problem.target_problem.targetproblem method)": [[29, "uo.target_problem.target_problem.TargetProblem.load_from_file"]], "name (uo.target_problem.target_problem.targetproblem property)": [[29, "uo.target_problem.target_problem.TargetProblem.name"]], "string_rep() (uo.target_problem.target_problem.targetproblem method)": [[29, "uo.target_problem.target_problem.TargetProblem.string_rep"]], "uo.target_problem": [[29, "module-uo.target_problem"]], "uo.target_problem.target_problem": [[29, "module-uo.target_problem.target_problem"]], "evaluationcachecontrolstatistics (class in uo.target_solution.evaluation_cache_control_statistics)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics"]], "qualityofsolution (class in uo.target_solution.target_solution)": [[30, "uo.target_solution.target_solution.QualityOfSolution"]], "targetsolution (class in uo.target_solution.target_solution)": [[30, "uo.target_solution.target_solution.TargetSolution"]], "cache (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.cache"]], "cache_hit_count (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.cache_hit_count"]], "cache_request_count (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.cache_request_count"]], "calculate_quality() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.calculate_quality"]], "calculate_quality_directly() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.calculate_quality_directly"]], "copy() (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics method)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.copy"]], "copy() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.copy"]], "copy_to() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.copy_to"]], "evaluate() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.evaluate"]], "fitness_value (uo.target_solution.target_solution.qualityofsolution attribute)": [[30, "uo.target_solution.target_solution.QualityOfSolution.fitness_value"]], "fitness_value (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.fitness_value"]], "increment_cache_hit_count() (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics method)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.increment_cache_hit_count"]], "increment_cache_request_count() (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics method)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.increment_cache_request_count"]], "init_from() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.init_from"]], "init_random() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.init_random"]], "is_caching (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.is_caching"]], "is_feasible (uo.target_solution.target_solution.qualityofsolution attribute)": [[30, "uo.target_solution.target_solution.QualityOfSolution.is_feasible"]], "is_feasible (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.is_feasible"]], "max_cache_size (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.max_cache_size"]], "name (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.name"]], "native_representation() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.native_representation"]], "objective_value (uo.target_solution.target_solution.qualityofsolution attribute)": [[30, "uo.target_solution.target_solution.QualityOfSolution.objective_value"]], "objective_value (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.objective_value"]], "representation (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.representation"]], "representation_distance() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.representation_distance"]], "representation_distance_directly() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.representation_distance_directly"]], "string_rep() (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics method)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.string_rep"]], "string_rep() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.string_rep"]], "string_representation() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.string_representation"]], "uo.target_solution": [[30, "module-uo.target_solution"]], "uo.target_solution.evaluation_cache_control_statistics": [[30, "module-uo.target_solution.evaluation_cache_control_statistics"]], "uo.target_solution.target_solution": [[30, "module-uo.target_solution.target_solution"]], "ensure_dir() (in module uo.utils.files)": [[31, "uo.utils.files.ensure_dir"]], "uo.utils": [[31, "module-uo.utils"]], "uo.utils.files": [[31, "module-uo.utils.files"]], "uo.utils.logger": [[31, "module-uo.utils.logger"]]}})
+Search.setIndex({"docnames": ["changelog", "contribution", "how_to_guides", "implemented_algorithms", "implemented_algorithms_exact", "implemented_algorithms_metaheuristic", "implemented_algorithms_total_enumeration", "implemented_algorithms_vns", "index", "modules", "opt", "opt.multi_objective", "opt.single_objective", "opt.single_objective.combinatorial", "opt.single_objective.combinatorial.traveling_salesperson_problem", "opt.single_objective.global", "opt.single_objective.teaching", "opt.single_objective.teaching.max_ones_problem", "problems_to_be_solved", "problems_to_be_solved_max_ones", "tutorials", "uo", "uo.algorithm", "uo.algorithm.exact", "uo.algorithm.exact.total_enumeration", "uo.algorithm.metaheuristic", "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic", "uo.algorithm.metaheuristic.genetic_algorithm", "uo.algorithm.metaheuristic.variable_neighborhood_search", "uo.target_problem", "uo.target_solution", "uo.utils"], "filenames": ["changelog.rst", "contribution.rst", "how_to_guides.rst", "implemented_algorithms.rst", "implemented_algorithms_exact.rst", "implemented_algorithms_metaheuristic.rst", "implemented_algorithms_total_enumeration.rst", "implemented_algorithms_vns.rst", "index.rst", "modules.rst", "opt.rst", "opt.multi_objective.rst", "opt.single_objective.rst", "opt.single_objective.combinatorial.rst", "opt.single_objective.combinatorial.traveling_salesperson_problem.rst", "opt.single_objective.global.rst", "opt.single_objective.teaching.rst", "opt.single_objective.teaching.max_ones_problem.rst", "problems_to_be_solved.rst", "problems_to_be_solved_max_ones.rst", "tutorials.rst", "uo.rst", "uo.algorithm.rst", "uo.algorithm.exact.rst", "uo.algorithm.exact.total_enumeration.rst", "uo.algorithm.metaheuristic.rst", "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic.rst", "uo.algorithm.metaheuristic.genetic_algorithm.rst", "uo.algorithm.metaheuristic.variable_neighborhood_search.rst", "uo.target_problem.rst", "uo.target_solution.rst", "uo.utils.rst"], "titles": ["Changelog", "How to Contribute", "How-to Guides", "Implemented algorithms", "Implemented exact algorithms", "Implemented metaheuristic algorithms", "Total Enumeration", "Variable Neighborhood Search", "Universal optimizer - Optimization algorithms library", "Library", "opt package", "opt.multi_objective package", "opt.single_objective package", "opt.single_objective.combinatorial package", "opt.single_objective.combinatorial.traveling_salesperson_problem package", "opt.single_objective.global package", "opt.single_objective.teaching package", "opt.single_objective.teaching.max_ones_problem package", "Problems to be solved", "Max Ones Problem", "Tutorials", "uo package", "uo.algorithm package", "uo.algorithm.exact package", "uo.algorithm.exact.total_enumeration package", "uo.algorithm.metaheuristic package", "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic package", "uo.algorithm.metaheuristic.genetic_algorithm package", "uo.algorithm.metaheuristic.variable_neighborhood_search package", "uo.target_problem package", "uo.target_solution package", "uo.utils package"], "terms": {"version": [0, 2], "0": [0, 17, 22, 24, 25, 28, 29, 30], "1": [0, 7, 17, 24, 25, 28, 30], "date": 0, "04": 0, "06": 0, "2023": 0, "descript": 0, "initi": [0, 1, 2, 7, 17, 22, 24, 28, 30], "setup": 0, "i": [0, 1, 2, 7, 8, 17, 19, 22, 24, 28, 30], "done": 0, "librari": [0, 1, 2], "structur": [0, 1, 3, 5, 8, 17, 25], "set": [0, 7, 17, 24, 29, 30], "2": [0, 7, 30], "26": 0, "metaheurist": [0, 1, 3, 7, 8, 21, 22], "algorithm": [0, 1, 9, 17, 21, 29], "vn": [0, 2, 7, 17, 24, 28], "ad": 0, "problem": [0, 1, 2, 7, 8, 17, 22, 24, 28, 29, 30], "maxon": 0, "4": 0, "01": 0, "07": 0, "document": [0, 1, 2, 7, 8], "support": [0, 7, 17], "should": [1, 22, 28, 31], "deriv": 1, "from": [1, 2, 7, 17, 28, 29, 30], "specifi": 1, "class": [1, 2, 7, 17, 19, 22, 24, 25, 28, 29, 30], "implement": [1, 8, 19, 28], "optim": [1, 7, 17, 22, 24, 25, 28], "uo": [1, 2, 8, 9], "place": [1, 29], "separ": [1, 22], "directori": [1, 2, 31], "within": [1, 2, 7, 22, 25, 28, 29], "exact": [1, 3, 8, 21, 22], "all": [1, 2, 7, 17, 24, 28], "program": [1, 17], "object": [1, 17, 22, 24, 29, 30], "function": [1, 17, 22, 31], "etc": [1, 7], "properli": 1, "us": [1, 2, 7, 8, 17, 24, 25, 28, 30], "system": [1, 31], "sphinx": [1, 2, 8], "restructuredtext": 1, "doc": [1, 2], "comment": 1, "code": [1, 2, 17, 28, 30], "have": [1, 17, 22], "web": 1, "page": [1, 8], "where": [1, 17, 22], "describ": [1, 22, 24, 25, 28, 29, 30], "There": [1, 7], "link": 1, "toward": 1, "explain": 1, "cover": 1, "unit": [1, 2, 8], "test": [1, 2, 8], "under": 1, "In": [1, 7], "order": 1, "do": 1, "so": [1, 22, 24], "unittest": [1, 2, 8], "exampl": [1, 2], "least": 1, "one": [1, 7, 22], "those": 1, "root": 1, "folder": [1, 2], "file": [1, 2, 9, 17, 21, 22, 29], "name": [1, 21, 22, 25, 29, 30], "py": 1, "vladimir": 1, "filipovi\u0107": 1, "vladofilipov": 1, "hotmail": 1, "com": 1, "variabl": [1, 3, 5, 8, 28], "neighborhood": [1, 3, 5, 8, 28], "search": [1, 3, 5, 8, 17, 28], "instal": 2, "provid": 2, "sourc": [2, 22], "poetri": [2, 8], "pip": 2, "check": [2, 22, 24], "successfulli": 2, "creat": [2, 24, 28], "virtual": 2, "environ": 2, "alreadi": [2, 17, 30], "same": [2, 7, 17, 22, 24, 25, 28, 29, 30], "command": [2, 17], "activ": 2, "deactiv": 2, "shell": 2, "project": [2, 8], "": [2, 17, 30], "packet": 2, "builder": 2, "depend": [2, 17], "execut": [2, 17, 22, 24, 25, 28], "develop": [2, 8], "applic": [2, 7], "strip": 2, "illustr": 2, "max": [2, 8, 17, 18], "Ones": [2, 8, 17, 18], "binari": [2, 17], "represent": [2, 17, 19, 21, 22, 24, 25, 28, 29, 30], "int": [2, 17, 19, 22, 24, 25, 28, 29, 30], "onli": [2, 7], "python": [2, 8], "example_vns_maxones_int_vns_exec": 2, "enhanc": 2, "both": [2, 7, 22], "app": 2, "example_vns_maxones_int_vns_exec_build": 2, "run": 2, "m": [2, 7], "build": 2, "apidoc": 2, "o": 2, "opt": [2, 8, 9], "chang": [2, 7], "current": [2, 7, 22, 24, 25, 28, 29, 30], "cd": 2, "clean": 2, "previous": 2, "html": 2, "make": 2, "total": [3, 4, 8, 24], "enumer": [3, 4, 8, 24], "basic": [3, 5, 8], "inform": [3, 5], "note": [3, 5, 8], "refer": [3, 5, 8], "denot": 7, "propos": 7, "mladenovi\u0107": 7, "hansen": 7, "mlahan1997": 7, "method": [7, 8], "solv": [7, 8, 17, 22, 24, 28, 30], "combinatori": [7, 10, 12], "global": [7, 10, 12], "It": 7, "explor": 7, "distant": 7, "incumb": 7, "solut": [7, 17, 19, 22, 24, 28, 30], "move": 7, "new": [7, 17, 22, 24, 25, 28, 29, 30], "an": 7, "improv": [7, 17, 28], "wa": 7, "made": 7, "The": [7, 17, 22, 24, 25, 28, 29, 30, 31], "local": [7, 17, 28], "appli": 7, "repeatedli": 7, "get": 7, "optima": 7, "systemat": 7, "two": [7, 17, 30], "phase": 7, "firstli": 7, "descent": 7, "find": 7, "optimum": 7, "final": 7, "perturb": 7, "out": 7, "correspond": 7, "vallei": 7, "ar": [7, 22, 30], "rapidli": 7, "increas": 7, "number": [7, 22, 24, 30], "pertain": 7, "mani": 7, "field": [7, 17, 21, 22, 24, 25, 28, 29, 30], "locat": 7, "theori": 7, "cluster": 7, "analysi": 7, "schedul": 7, "vehicl": 7, "rout": 7, "network": 7, "design": 7, "lot": 7, "size": [7, 30], "artifici": 7, "intellig": 7, "engin": 7, "pool": 7, "biologi": 7, "phylogeni": 7, "reliabl": 7, "geometri": 7, "telecommun": 7, "sever": [7, 28], "book": 7, "import": 7, "understand": 7, "like": 7, "handbook": 7, "genpot2010": 7, "glokoc2003": 7, "methodologi": 7, "burken2005": 7, "accord": 7, "autor": 7, "perform": 7, "procedur": [7, 17, 28], "minima": 7, "escap": 7, "which": [7, 17, 22, 30], "contain": [7, 8, 17, 28, 31], "them": 7, "built": 7, "upon": 7, "follow": [7, 22], "percept": 7, "A": 7, "minimum": 7, "respect": 7, "necessarili": 7, "anoth": 7, "possibl": 7, "For": 7, "rel": 7, "close": 7, "each": 7, "other": 7, "main": [7, 16, 17, 25, 28], "step": [7, 22], "through": 7, "choos": 7, "x": [7, 17, 24, 25, 28], "discov": 7, "direct": 7, "given": [7, 19], "k": [7, 17, 28], "proceed": 7, "f": 7, "If": 7, "heurist": 7, "stop": 7, "otherwis": 7, "iter": [7, 21, 22, 24, 25, 28], "usual": 7, "highest": 7, "also": 7, "relat": 7, "best": [7, 17, 22, 28], "shake": [7, 16, 17, 25, 28], "typic": 7, "involv": 7, "random": [7, 16, 17, 22, 24, 25, 28, 30], "extract": 7, "feasibl": [7, 17, 30], "compar": 7, "valu": [7, 17, 22, 24, 25, 28, 29, 30], "obtain": [7, 17, 22, 30], "return": [7, 17, 22, 24, 25, 28, 29, 30], "its": [7, 17, 30], "updat": [7, 22], "next": [7, 24], "consid": 7, "depth": 7, "variou": [7, 8, 17], "variant": [7, 17, 28], "vnsoptim": [7, 25, 28], "n": [7, 17], "p": 7, "1997": 7, "comput": 7, "oper": [7, 22, 24], "research": 7, "24": 7, "11": 7, "1097": 7, "1100": 7, "citeseerx": 7, "10": 7, "800": 7, "1797": 7, "doi": 7, "1016": 7, "s0305": 7, "0548": 7, "97": 7, "00031": 7, "gendreau": 7, "potvin": 7, "j": 7, "y": 7, "2010": 7, "springer": 7, "glover": 7, "kochenberg": 7, "g": [7, 30], "2003": 7, "kluwer": 7, "academ": 7, "publish": 7, "burk": 7, "ek": 7, "kendal": 7, "2005": 7, "edmund": 7, "graham": 7, "ed": 7, "introductori": 7, "tutori": 7, "decis": 7, "techniqu": [7, 17], "1007": 7, "978": 7, "4614": 7, "6940": 7, "7": 7, "isbn": 7, "6939": 7, "languag": 8, "tool": 8, "manag": 8, "gener": [8, 22, 24, 28, 30], "packag": [8, 9], "subpackag": [8, 9], "target_problem": [8, 9, 21, 22, 24, 25, 28, 30], "target_solut": [8, 9, 21, 24], "util": [8, 9, 21], "modul": [8, 9], "content": [8, 9], "multi_object": [8, 9, 10], "single_object": [8, 9, 10], "contributor": 8, "index": 8, "submodul": [9, 12, 16, 21], "output_control": [9, 21, 24, 25, 28], "evaluation_cache_control_statist": [9, 21], "logger": [9, 21], "trivial": [10, 12], "traveling_salesperson_problem": [12, 13], "max_ones_problem": [12, 16], "command_lin": [12, 16], "max_ones_problem_binary_bit_array_solut": [12, 16], "max_ones_problem_binary_bit_array_solution_vns_support": [12, 16], "max_ones_problem_binary_int_solut": [12, 16], "max_ones_problem_binary_int_solution_vns_support": [12, 16], "solver": [12, 16], "parse_argu": [16, 17], "maxonesproblem": [16, 17, 19], "copi": [16, 17, 21, 22, 24, 25, 28, 29, 30], "load_from_fil": [16, 17, 21, 29], "string_rep": [16, 17, 21, 22, 24, 25, 28, 29, 30], "maxonesproblembinarybitarraysolut": [16, 17, 19], "calculate_quality_directli": [16, 17, 21, 30], "copy_to": [16, 17, 21, 30], "init_from": [16, 17, 21, 30], "init_random": [16, 17, 21, 30], "native_represent": [16, 17, 21, 30], "representation_distance_directli": [16, 17, 21, 30], "string_represent": [16, 17, 21, 22, 30], "maxonesproblembinarybitarraysolutionvnssupport": [16, 17], "local_search_best_improv": [16, 17, 25, 28], "local_search_first_improv": [16, 17, 25, 28], "maxonesproblembinaryintsolut": [16, 17, 19], "maxonesproblembinaryintsolutionvnssupport": [16, 17], "paramet": [17, 22, 24, 25, 28, 29, 30], "ones": 17, "pars": 17, "repres": [17, 19, 24, 25, 28, 30], "file_path": [17, 21, 29], "str": [17, 22, 24, 25, 28, 29, 30, 31], "none": [17, 22, 24, 25, 28, 29, 30, 31], "dim": 17, "base": [17, 22, 24, 25, 28, 29, 30], "targetproblem": [17, 21, 22, 24, 25, 28, 29, 30], "instanc": [17, 22, 24, 25, 28, 29, 30], "properti": [17, 22, 24, 25, 28, 29, 30], "type": [17, 22, 24, 25, 28, 29, 30], "data_format": 17, "txt": 17, "read": [17, 29], "target": [17, 22, 29, 30], "data": [17, 22, 29], "format": 17, "delimit": [17, 22, 24, 25, 28, 29, 30], "indent": [17, 22, 24, 25, 28, 29, 30], "indentation_symbol": [17, 22, 24, 25, 28, 29, 30], "group_start": [17, 22, 24, 25, 28, 29, 30], "group_end": [17, 22, 24, 25, 28, 29, 30], "string": [17, 22, 24, 25, 28, 29, 30], "maxoneproblem": 17, "between": [17, 22, 24, 25, 28, 29, 30], "option": [17, 22, 24, 25, 28, 29, 30], "default": [17, 22, 24, 25, 28, 29, 30], "level": [17, 22, 24, 25, 28, 29, 30], "symbol": [17, 22, 24, 25, 28, 29, 30], "group": [17, 22, 24, 25, 28, 29, 30], "start": [17, 22, 24, 25, 28, 29, 30], "end": [17, 22, 24, 25, 28, 29, 30], "control": [17, 22, 24, 25, 28, 29, 30], "output": [17, 22, 24, 25, 28, 29, 30], "bitarrai": [17, 19], "ha": 17, "been": 17, "random_se": [17, 22, 25, 28, 30], "evaluation_cache_is_us": [17, 30], "bool": [17, 22, 24, 28, 29, 30], "fals": [17, 22, 30], "evaluation_cache_max_s": [17, 30], "distance_calculation_cache_is_us": [17, 30], "distance_calculation_cache_max_s": [17, 30], "targetsolut": [17, 21, 22, 24, 28, 30], "qualityofsolut": [17, 21, 30], "fit": [17, 22, 30], "calcul": [17, 25, 30], "nativ": [17, 30], "whose": [17, 31], "destin": [17, 30], "exist": [17, 30, 31], "ste": [17, 30], "representation_str": [17, 30], "solution_code_2": 17, "float": [17, 25, 30], "distanc": [17, 30], "determin": [17, 22, 30], "solution_code_1": 17, "first": [17, 22, 28, 30], "second": [17, 22, 25, 30], "interv": [17, 24, 25, 28], "part": [17, 28], "problemsolutionvnssupport": [17, 25, 28], "result": [17, 28], "doe": [17, 28], "differ": [17, 28], "more": [17, 28], "than": [17, 22, 28], "insid": [17, 28], "shakingpoint": 17, "success": [17, 28], "integ": 17, "directli": [17, 30], "representation_1": [17, 30], "r_co": [17, 24, 28, 30], "representation_2": [17, 30], "thi": [17, 22, 24, 25], "line": 17, "solution_code_distance_cache_control_statist": [21, 22], "best_solut": [21, 22], "copy_to_best_solut": [21, 22], "evalu": [21, 22, 24, 30], "execution_end": [21, 22], "execution_start": [21, 22], "init": [21, 22, 24, 25, 28], "is_first_solution_bett": [21, 22], "write_output_headers_if_need": [21, 22], "write_output_values_if_need": [21, 22], "outputcontrol": [21, 22, 24, 25, 28], "fields_definit": [21, 22], "fields_head": [21, 22], "moment": [21, 22], "output_fil": [21, 22], "write_after_algorithm": [21, 22], "write_after_evalu": [21, 22], "write_after_iter": [21, 22], "write_after_step_in_iter": [21, 22], "write_before_algorithm": [21, 22], "write_before_evalu": [21, 22], "write_before_iter": [21, 22], "write_before_step_in_iter": [21, 22], "write_to_output": [21, 22], "dimens": [21, 29], "is_minim": [21, 29], "evaluationcachecontrolstatist": [21, 30], "cach": [21, 30], "cache_hit_count": [21, 30], "cache_request_count": [21, 30], "increment_cache_hit_count": [21, 30], "increment_cache_request_count": [21, 30], "is_cach": [21, 30], "max_cache_s": [21, 30], "fitness_valu": [21, 22, 30], "is_feas": [21, 22, 30], "objective_valu": [21, 22, 30], "calculate_qu": [21, 30], "representation_dist": [21, 30], "ensure_dir": [21, 31], "electro_magnetism_like_metaheusrist": [22, 25], "genetic_algorithm": [22, 25], "variable_neighborhood_search": [22, 25], "problem_solution_vns_support": [22, 25], "vns_optim": [22, 25], "additional_statistics_control": [22, 25, 28], "elapsed_second": [22, 25], "finish_control": [22, 25, 28], "main_loop": [22, 25], "main_loop_iter": [22, 25, 28], "getter": [22, 24, 25, 28, 29, 30], "dure": [22, 24, 25, 30], "far": 22, "abstract": [22, 24, 25, 28, 29, 30], "argument": 22, "becom": 22, "info": 22, "about": 22, "time": [22, 25], "when": 22, "cope": 22, "datetim": 22, "sol1": 22, "sol2": 22, "better": 22, "true": 22, "wors": 22, "equal": 22, "write": 22, "header": 22, "necessari": 22, "step_nam": 22, "step_name_valu": 22, "written": 22, "after_algorithm": 22, "before_algorithm": 22, "after_iter": 22, "before_iter": 22, "after_evalu": 22, "before_evalu": 22, "after_step_in_iter": 22, "before_step_in_iter": 22, "what": 22, "instead": 22, "textiowrapp": 22, "comma": 22, "list": [22, 28, 30], "definit": 22, "head": 22, "after": 22, "prior": 22, "problemsolutiontesupport": 24, "can_progress": 24, "process": 24, "indic": 24, "overall_number_of_evalu": 24, "overal": [24, 30], "requir": 24, "finish": [24, 25], "progress": 24, "intern": 24, "counter": 24, "configur": 24, "taken": 24, "consider": 24, "state": 24, "reflect": 24, "reset": 24, "over": 24, "totalenumer": 24, "teoptim": 24, "initial_solut": [24, 28], "current_solut": 24, "subtyp": 24, "classmethod": [24, 28], "from_construction_tupl": [24, 25, 28], "construction_tupl": [24, 28], "teoptimizerconstructionparamet": 24, "addit": [24, 28], "constructor": [24, 28], "tupl": [24, 28, 30], "te": 24, "totalenumerationconstructorparamet": 24, "k_max": [25, 28], "k_min": [25, 28], "finishcontrol": [25, 28], "additionalstatisticscontrol": [25, 28], "keep": 25, "statist": [25, 30], "elaps": 25, "criteria": 25, "loop": [25, 28], "One": [25, 28], "seed": 25, "solution_represent": 28, "suppli": 28, "collect": 28, "shaken": 28, "local_search_typ": 28, "singlesolutionmetaheurist": 28, "encapsul": 28, "vnsoptimizerconstructionparamet": 28, "path": [29, 31], "minim": 29, "take": 29, "intrins": 29, "data_represent": 29, "dict": 30, "hit": 30, "increment": 30, "maximum": 30, "unlimit": 30, "alia": 30, "consult": 30, "e": 30, "deal": 31, "path_to_dir": 31, "ensur": 31, "specif": 31, "param": 31, "log": 31}, "objects": {"": [[10, 0, 0, "-", "opt"], [21, 0, 0, "-", "uo"]], "opt": [[11, 0, 0, "-", "multi_objective"], [12, 0, 0, "-", "single_objective"]], "opt.single_objective": [[13, 0, 0, "-", "combinatorial"], [15, 0, 0, "-", "global"], [16, 0, 0, "-", "trivial"]], "opt.single_objective.combinatorial": [[14, 0, 0, "-", "traveling_salesperson_problem"]], "opt.single_objective.teaching": [[17, 0, 0, "-", "max_ones_problem"]], "opt.single_objective.teaching.max_ones_problem": [[17, 0, 0, "-", "command_line"], [17, 0, 0, "-", "max_ones_problem"], [17, 0, 0, "-", "max_ones_problem_binary_bit_array_solution"], [17, 0, 0, "-", "max_ones_problem_binary_bit_array_solution_vns_support"], [17, 0, 0, "-", "max_ones_problem_binary_int_solution"], [17, 0, 0, "-", "max_ones_problem_binary_int_solution_vns_support"], [17, 0, 0, "-", "solver"]], "opt.single_objective.teaching.max_ones_problem.command_line": [[17, 1, 1, "", "parse_arguments"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem": [[17, 2, 1, "", "MaxOnesProblem"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem.MaxOnesProblem": [[17, 3, 1, "", "copy"], [17, 3, 1, "", "load_from_file"], [17, 3, 1, "", "string_rep"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution": [[17, 2, 1, "", "MaxOnesProblemBinaryBitArraySolution"], [17, 1, 1, "", "random"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution": [[17, 3, 1, "", "calculate_quality_directly"], [17, 3, 1, "", "copy"], [17, 3, 1, "", "copy_to"], [17, 3, 1, "", "init_from"], [17, 3, 1, "", "init_random"], [17, 3, 1, "", "native_representation"], [17, 3, 1, "", "representation_distance_directly"], [17, 3, 1, "", "string_rep"], [17, 3, 1, "", "string_representation"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support": [[17, 2, 1, "", "MaxOnesProblemBinaryBitArraySolutionVnsSupport"], [17, 1, 1, "", "random"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport": [[17, 3, 1, "", "copy"], [17, 3, 1, "", "local_search_best_improvement"], [17, 3, 1, "", "local_search_first_improvement"], [17, 3, 1, "", "shaking"], [17, 3, 1, "", "string_rep"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution": [[17, 2, 1, "", "MaxOnesProblemBinaryIntSolution"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution": [[17, 3, 1, "", "calculate_quality_directly"], [17, 3, 1, "", "copy"], [17, 3, 1, "", "copy_to"], [17, 3, 1, "", "init_from"], [17, 3, 1, "", "init_random"], [17, 3, 1, "", "native_representation"], [17, 3, 1, "", "representation_distance_directly"], [17, 3, 1, "", "string_rep"], [17, 3, 1, "", "string_representation"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support": [[17, 2, 1, "", "MaxOnesProblemBinaryIntSolutionVnsSupport"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport": [[17, 3, 1, "", "copy"], [17, 3, 1, "", "local_search_best_improvement"], [17, 3, 1, "", "local_search_first_improvement"], [17, 3, 1, "", "shaking"], [17, 3, 1, "", "string_rep"]], "opt.single_objective.teaching.max_ones_problem.solver": [[17, 1, 1, "", "main"]], "uo": [[22, 0, 0, "-", "algorithm"], [29, 0, 0, "-", "target_problem"], [30, 0, 0, "-", "target_solution"], [31, 0, 0, "-", "utils"]], "uo.algorithm": [[22, 0, 0, "-", "algorithm"], [23, 0, 0, "-", "exact"], [25, 0, 0, "-", "metaheuristic"], [22, 0, 0, "-", "output_control"]], "uo.algorithm.algorithm": [[22, 2, 1, "", "Algorithm"]], "uo.algorithm.algorithm.Algorithm": [[22, 4, 1, "", "best_solution"], [22, 3, 1, "", "copy"], [22, 3, 1, "", "copy_to_best_solution"], [22, 4, 1, "", "evaluation"], [22, 4, 1, "", "execution_ended"], [22, 4, 1, "", "execution_started"], [22, 3, 1, "", "init"], [22, 3, 1, "", "is_first_solution_better"], [22, 4, 1, "", "iteration"], [22, 4, 1, "", "name"], [22, 4, 1, "", "output_control"], [22, 3, 1, "", "string_rep"], [22, 4, 1, "", "target_problem"], [22, 3, 1, "", "write_output_headers_if_needed"], [22, 3, 1, "", "write_output_values_if_needed"]], "uo.algorithm.exact": [[24, 0, 0, "-", "total_enumeration"]], "uo.algorithm.exact.total_enumeration": [[24, 0, 0, "-", "problem_solution_te_support"], [24, 0, 0, "-", "te_optimizer"], [24, 0, 0, "-", "te_optimizer_constructor_parameters"]], "uo.algorithm.exact.total_enumeration.problem_solution_te_support": [[24, 2, 1, "", "ProblemSolutionTeSupport"]], "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport": [[24, 3, 1, "", "can_progress"], [24, 3, 1, "", "overall_number_of_evaluations"], [24, 3, 1, "", "progress"], [24, 3, 1, "", "reset"]], "uo.algorithm.exact.total_enumeration.te_optimizer": [[24, 2, 1, "", "TeOptimizer"], [24, 1, 1, "", "random"]], "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer": [[24, 3, 1, "", "copy"], [24, 4, 1, "", "current_solution"], [24, 3, 1, "", "from_construction_tuple"], [24, 3, 1, "", "init"], [24, 4, 1, "", "iteration"], [24, 3, 1, "", "optimize"], [24, 3, 1, "", "string_rep"]], "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters": [[24, 2, 1, "", "TeOptimizerConstructionParameters"]], "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters": [[24, 5, 1, "", "initial_solution"], [24, 5, 1, "", "output_control"], [24, 5, 1, "", "problem_solution_te_support"], [24, 5, 1, "", "target_problem"]], "uo.algorithm.metaheuristic": [[26, 0, 0, "-", "electro_magnetism_like_metaheusristic"], [27, 0, 0, "-", "genetic_algorithm"], [25, 0, 0, "-", "metaheuristic"], [28, 0, 0, "-", "variable_neighborhood_search"]], "uo.algorithm.metaheuristic.metaheuristic": [[25, 2, 1, "", "Metaheuristic"], [25, 1, 1, "", "random"]], "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic": [[25, 4, 1, "", "additional_statistics_control"], [25, 3, 1, "", "copy"], [25, 3, 1, "", "elapsed_seconds"], [25, 4, 1, "", "finish_control"], [25, 3, 1, "", "main_loop"], [25, 3, 1, "", "main_loop_iteration"], [25, 3, 1, "", "optimize"], [25, 4, 1, "", "random_seed"], [25, 3, 1, "", "string_rep"]], "uo.algorithm.metaheuristic.variable_neighborhood_search": [[28, 0, 0, "-", "problem_solution_vns_support"], [28, 0, 0, "-", "vns_optimizer"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support": [[28, 2, 1, "", "ProblemSolutionVnsSupport"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport": [[28, 3, 1, "", "local_search_best_improvement"], [28, 3, 1, "", "local_search_first_improvement"], [28, 3, 1, "", "shaking"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer": [[28, 2, 1, "", "VnsOptimizer"], [28, 1, 1, "", "random"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer": [[28, 3, 1, "", "copy"], [28, 3, 1, "", "from_construction_tuple"], [28, 3, 1, "", "init"], [28, 4, 1, "", "k_max"], [28, 4, 1, "", "k_min"], [28, 3, 1, "", "main_loop_iteration"], [28, 3, 1, "", "string_rep"]], "uo.algorithm.output_control": [[22, 2, 1, "", "OutputControl"]], "uo.algorithm.output_control.OutputControl": [[22, 4, 1, "", "fields"], [22, 4, 1, "", "fields_definitions"], [22, 4, 1, "", "fields_headings"], [22, 4, 1, "", "moments"], [22, 4, 1, "", "output_file"], [22, 3, 1, "", "string_rep"], [22, 4, 1, "", "write_after_algorithm"], [22, 4, 1, "", "write_after_evaluation"], [22, 4, 1, "", "write_after_iteration"], [22, 4, 1, "", "write_after_step_in_iteration"], [22, 4, 1, "", "write_before_algorithm"], [22, 4, 1, "", "write_before_evaluation"], [22, 4, 1, "", "write_before_iteration"], [22, 4, 1, "", "write_before_step_in_iteration"], [22, 4, 1, "", "write_to_output"]], "uo.target_problem": [[29, 0, 0, "-", "target_problem"]], "uo.target_problem.target_problem": [[29, 2, 1, "", "TargetProblem"]], "uo.target_problem.target_problem.TargetProblem": [[29, 3, 1, "", "copy"], [29, 4, 1, "", "dimension"], [29, 4, 1, "", "file_path"], [29, 4, 1, "", "is_minimization"], [29, 3, 1, "", "load_from_file"], [29, 4, 1, "", "name"], [29, 3, 1, "", "string_rep"]], "uo.target_solution": [[30, 0, 0, "-", "evaluation_cache_control_statistics"], [30, 0, 0, "-", "target_solution"]], "uo.target_solution.evaluation_cache_control_statistics": [[30, 2, 1, "", "EvaluationCacheControlStatistics"]], "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics": [[30, 4, 1, "", "cache"], [30, 4, 1, "", "cache_hit_count"], [30, 4, 1, "", "cache_request_count"], [30, 3, 1, "", "copy"], [30, 3, 1, "", "increment_cache_hit_count"], [30, 3, 1, "", "increment_cache_request_count"], [30, 4, 1, "", "is_caching"], [30, 4, 1, "", "max_cache_size"], [30, 3, 1, "", "string_rep"]], "uo.target_solution.target_solution": [[30, 2, 1, "", "QualityOfSolution"], [30, 2, 1, "", "TargetSolution"]], "uo.target_solution.target_solution.QualityOfSolution": [[30, 5, 1, "", "fitness_value"], [30, 5, 1, "", "is_feasible"], [30, 5, 1, "", "objective_value"]], "uo.target_solution.target_solution.TargetSolution": [[30, 3, 1, "", "calculate_quality"], [30, 3, 1, "", "calculate_quality_directly"], [30, 3, 1, "", "copy"], [30, 3, 1, "", "copy_to"], [30, 3, 1, "", "evaluate"], [30, 4, 1, "", "fitness_value"], [30, 3, 1, "", "init_from"], [30, 3, 1, "", "init_random"], [30, 4, 1, "", "is_feasible"], [30, 4, 1, "", "name"], [30, 3, 1, "", "native_representation"], [30, 4, 1, "", "objective_value"], [30, 4, 1, "", "representation"], [30, 3, 1, "", "representation_distance"], [30, 3, 1, "", "representation_distance_directly"], [30, 3, 1, "", "string_rep"], [30, 3, 1, "", "string_representation"]], "uo.utils": [[31, 0, 0, "-", "files"], [31, 0, 0, "-", "logger"]], "uo.utils.files": [[31, 1, 1, "", "ensure_dir"]]}, "objtypes": {"0": "py:module", "1": "py:function", "2": "py:class", "3": "py:method", "4": "py:property", "5": "py:attribute"}, "objnames": {"0": ["py", "module", "Python module"], "1": ["py", "function", "Python function"], "2": ["py", "class", "Python class"], "3": ["py", "method", "Python method"], "4": ["py", "property", "Python property"], "5": ["py", "attribute", "Python attribute"]}, "titleterms": {"changelog": [0, 8], "how": [1, 2, 8], "contribut": [1, 8], "contributor": 1, "guid": [2, 8], "implement": [3, 4, 5, 7], "algorithm": [3, 4, 5, 7, 8, 22, 23, 24, 25, 26, 27, 28], "exact": [4, 23, 24], "metaheurist": [5, 25, 26, 27, 28], "total": 6, "enumer": 6, "variabl": 7, "neighborhood": 7, "search": 7, "basic": 7, "inform": [7, 8], "structur": 7, "note": 7, "refer": 7, "univers": 8, "optim": [8, 9], "librari": [8, 9], "tutori": [8, 20], "To": 8, "In": 8, "depth": 8, "program": 8, "code": 8, "indic": 8, "tabl": 8, "opt": [10, 11, 12, 13, 14, 15, 16, 17], "packag": [10, 11, 12, 13, 14, 15, 16, 17, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31], "subpackag": [10, 12, 13, 16, 21, 22, 25], "modul": [10, 11, 12, 13, 14, 15, 16, 17, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31], "content": [10, 11, 12, 13, 14, 15, 16, 17, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31], "multi_object": 11, "single_object": [12, 13, 14, 15, 16, 17], "combinatori": [13, 14], "traveling_salesperson_problem": 14, "global": 15, "trivial": [16, 17], "max_ones_problem": 17, "submodul": [17, 22, 24, 25, 28, 29, 30, 31], "command_lin": 17, "max_ones_problem_binary_bit_array_solut": 17, "max_ones_problem_binary_bit_array_solution_vns_support": 17, "max_ones_problem_binary_int_solut": 17, "max_ones_problem_binary_int_solution_vns_support": 17, "solver": 17, "problem": [18, 19], "solv": 18, "max": 19, "Ones": 19, "uo": [21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31], "output_control": 22, "total_enumer": 24, "problem_solution_te_support": 24, "te_optim": 24, "te_optimizer_constructor_paramet": 24, "solution_code_distance_cache_control_statist": 25, "electro_magnetism_like_metaheusrist": 26, "genetic_algorithm": 27, "variable_neighborhood_search": 28, "problem_solution_vns_support": 28, "vns_optim": 28, "target_problem": 29, "target_solut": 30, "evaluation_cache_control_statist": 30, "util": 31, "file": 31, "logger": 31}, "envversion": {"sphinx.domains.c": 3, "sphinx.domains.changeset": 1, "sphinx.domains.citation": 1, "sphinx.domains.cpp": 9, "sphinx.domains.index": 1, "sphinx.domains.javascript": 3, "sphinx.domains.math": 2, "sphinx.domains.python": 4, "sphinx.domains.rst": 2, "sphinx.domains.std": 2, "sphinx.ext.intersphinx": 1, "sphinx": 60}, "alltitles": {"Changelog": [[0, "changelog"]], "How to Contribute": [[1, "how-to-contribute"]], "Contributors": [[1, "contributors"]], "How-to Guides": [[2, "how-to-guides"]], "Implemented algorithms": [[3, "implemented-algorithms"], [3, null]], "Implemented exact algorithms": [[4, "implemented-exact-algorithms"], [4, null]], "Implemented metaheuristic algorithms": [[5, "implemented-metaheuristic-algorithms"], [5, null]], "Total Enumeration": [[6, "total-enumeration"]], "Variable Neighborhood Search": [[7, "variable-neighborhood-search"]], "Basic information": [[7, "basic-information"]], "Structure of the algorithm": [[7, "structure-of-the-algorithm"]], "Implementation notes": [[7, "implementation-notes"]], "References": [[7, "references"]], "Universal optimizer - Optimization algorithms library": [[8, "universal-optimizer-optimization-algorithms-library"]], "Tutorials and How-To Guides": [[8, null]], "In-depth Information": [[8, null]], "Program code": [[8, null]], "Changelog and Contribution": [[8, null]], "Indices and tables": [[8, "indices-and-tables"]], "Library": [[9, "library"]], "Optimization": [[9, "optimization"]], "opt package": [[10, "opt-package"]], "Subpackages": [[10, "subpackages"], [12, "subpackages"], [13, "subpackages"], [16, "subpackages"], [21, "subpackages"], [22, "subpackages"], [25, "subpackages"]], "Module contents": [[10, "module-opt"], [11, "module-opt.multi_objective"], [12, "module-opt.single_objective"], [13, "module-opt.single_objective.combinatorial"], [14, "module-opt.single_objective.combinatorial.traveling_salesperson_problem"], [15, "module-opt.single_objective.global"], [16, "module-opt.single_objective.teaching"], [17, "module-opt.single_objective.teaching.max_ones_problem"], [21, "module-uo"], [22, "module-uo.algorithm"], [23, "module-uo.algorithm.exact"], [24, "module-uo.algorithm.exact.total_enumeration"], [25, "module-uo.algorithm.metaheuristic"], [26, "module-uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic"], [27, "module-uo.algorithm.metaheuristic.genetic_algorithm"], [28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search"], [29, "module-uo.target_problem"], [30, "module-uo.target_solution"], [31, "module-uo.utils"]], "opt.multi_objective package": [[11, "opt-multi-objective-package"]], "opt.single_objective package": [[12, "opt-single-objective-package"]], "opt.single_objective.combinatorial package": [[13, "opt-single-objective-combinatorial-package"]], "opt.single_objective.combinatorial.traveling_salesperson_problem package": [[14, "opt-single-objective-combinatorial-traveling-salesperson-problem-package"]], "opt.single_objective.global package": [[15, "opt-single-objective-global-package"]], "opt.single_objective.teaching package": [[16, "opt-single-objective-trivial-package"]], "opt.single_objective.teaching.max_ones_problem package": [[17, "opt-single-objective-trivial-max-ones-problem-package"]], "Submodules": [[17, "submodules"], [22, "submodules"], [24, "submodules"], [25, "submodules"], [28, "submodules"], [29, "submodules"], [30, "submodules"], [31, "submodules"]], "opt.single_objective.teaching.max_ones_problem.command_line module": [[17, "module-opt.single_objective.teaching.max_ones_problem.command_line"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem module": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution module": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support module": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution module": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support module": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support"]], "opt.single_objective.teaching.max_ones_problem.solver module": [[17, "module-opt.single_objective.teaching.max_ones_problem.solver"]], "Problems to be solved": [[18, "problems-to-be-solved"]], "Problems": [[18, null]], "Max Ones Problem": [[19, "max-ones-problem"]], "Tutorials": [[20, "tutorials"]], "uo package": [[21, "uo-package"]], "uo.algorithm package": [[22, "uo-algorithm-package"]], "uo.algorithm.algorithm module": [[22, "module-uo.algorithm.algorithm"]], "uo.algorithm.output_control module": [[22, "module-uo.algorithm.output_control"]], "uo.algorithm.exact package": [[23, "uo-algorithm-exact-package"]], "uo.algorithm.exact.total_enumeration package": [[24, "uo-algorithm-exact-total-enumeration-package"]], "uo.algorithm.exact.total_enumeration.problem_solution_te_support module": [[24, "module-uo.algorithm.exact.total_enumeration.problem_solution_te_support"]], "uo.algorithm.exact.total_enumeration.te_optimizer module": [[24, "module-uo.algorithm.exact.total_enumeration.te_optimizer"]], "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters module": [[24, "module-uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters"]], "uo.algorithm.metaheuristic package": [[25, "uo-algorithm-metaheuristic-package"]], "uo.algorithm.metaheuristic.metaheuristic module": [[25, "module-uo.algorithm.metaheuristic.metaheuristic"]], "uo.algorithm.metaheuristic.solution_code_distance_cache_control_statistics module": [[25, "uo-algorithm-metaheuristic-solution-code-distance-cache-control-statistics-module"]], "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic package": [[26, "uo-algorithm-metaheuristic-electro-magnetism-like-metaheusristic-package"]], "uo.algorithm.metaheuristic.genetic_algorithm package": [[27, "uo-algorithm-metaheuristic-genetic-algorithm-package"]], "uo.algorithm.metaheuristic.variable_neighborhood_search package": [[28, "uo-algorithm-metaheuristic-variable-neighborhood-search-package"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support module": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer module": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer"]], "uo.target_problem package": [[29, "uo-target-problem-package"]], "uo.target_problem.target_problem module": [[29, "module-uo.target_problem.target_problem"]], "uo.target_solution package": [[30, "uo-target-solution-package"]], "uo.target_solution.evaluation_cache_control_statistics module": [[30, "module-uo.target_solution.evaluation_cache_control_statistics"]], "uo.target_solution.target_solution module": [[30, "module-uo.target_solution.target_solution"]], "uo.utils package": [[31, "uo-utils-package"]], "uo.utils.files module": [[31, "module-uo.utils.files"]], "uo.utils.logger module": [[31, "module-uo.utils.logger"]]}, "indexentries": {"module": [[10, "module-opt"], [11, "module-opt.multi_objective"], [12, "module-opt.single_objective"], [13, "module-opt.single_objective.combinatorial"], [14, "module-opt.single_objective.combinatorial.traveling_salesperson_problem"], [15, "module-opt.single_objective.global"], [16, "module-opt.single_objective.teaching"], [17, "module-opt.single_objective.teaching.max_ones_problem"], [17, "module-opt.single_objective.teaching.max_ones_problem.command_line"], [17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem"], [17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution"], [17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support"], [17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution"], [17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support"], [17, "module-opt.single_objective.teaching.max_ones_problem.solver"], [21, "module-uo"], [22, "module-uo.algorithm"], [22, "module-uo.algorithm.algorithm"], [22, "module-uo.algorithm.output_control"], [23, "module-uo.algorithm.exact"], [24, "module-uo.algorithm.exact.total_enumeration"], [24, "module-uo.algorithm.exact.total_enumeration.problem_solution_te_support"], [24, "module-uo.algorithm.exact.total_enumeration.te_optimizer"], [24, "module-uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters"], [25, "module-uo.algorithm.metaheuristic"], [25, "module-uo.algorithm.metaheuristic.metaheuristic"], [26, "module-uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic"], [27, "module-uo.algorithm.metaheuristic.genetic_algorithm"], [28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search"], [28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support"], [28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer"], [29, "module-uo.target_problem"], [29, "module-uo.target_problem.target_problem"], [30, "module-uo.target_solution"], [30, "module-uo.target_solution.evaluation_cache_control_statistics"], [30, "module-uo.target_solution.target_solution"], [31, "module-uo.utils"], [31, "module-uo.utils.files"], [31, "module-uo.utils.logger"]], "opt": [[10, "module-opt"]], "opt.multi_objective": [[11, "module-opt.multi_objective"]], "opt.single_objective": [[12, "module-opt.single_objective"]], "opt.single_objective.combinatorial": [[13, "module-opt.single_objective.combinatorial"]], "opt.single_objective.combinatorial.traveling_salesperson_problem": [[14, "module-opt.single_objective.combinatorial.traveling_salesperson_problem"]], "opt.single_objective.global": [[15, "module-opt.single_objective.global"]], "opt.single_objective.teaching": [[16, "module-opt.single_objective.teaching"]], "maxonesproblem (class in opt.single_objective.teaching.max_ones_problem.max_ones_problem)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem.MaxOnesProblem"]], "maxonesproblembinarybitarraysolution (class in opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution"]], "maxonesproblembinarybitarraysolutionvnssupport (class in opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport"]], "maxonesproblembinaryintsolution (class in opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution"]], "maxonesproblembinaryintsolutionvnssupport (class in opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport"]], "calculate_quality_directly() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.calculate_quality_directly"]], "calculate_quality_directly() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.calculate_quality_directly"]], "copy() (opt.single_objective.teaching.max_ones_problem.max_ones_problem.maxonesproblem method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem.MaxOnesProblem.copy"]], "copy() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.copy"]], "copy() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.copy"]], "copy() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.copy"]], "copy() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.copy"]], "copy_to() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.copy_to"]], "copy_to() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.copy_to"]], "init_from() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.init_from"]], "init_from() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.init_from"]], "init_random() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.init_random"]], "init_random() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.init_random"]], "load_from_file() (opt.single_objective.teaching.max_ones_problem.max_ones_problem.maxonesproblem method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem.MaxOnesProblem.load_from_file"]], "local_search_best_improvement() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.local_search_best_improvement"]], "local_search_best_improvement() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.local_search_best_improvement"]], "local_search_first_improvement() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.local_search_first_improvement"]], "local_search_first_improvement() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.local_search_first_improvement"]], "main() (in module opt.single_objective.teaching.max_ones_problem.solver)": [[17, "opt.single_objective.teaching.max_ones_problem.solver.main"]], "native_representation() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.native_representation"]], "native_representation() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.native_representation"]], "opt.single_objective.teaching.max_ones_problem": [[17, "module-opt.single_objective.teaching.max_ones_problem"]], "opt.single_objective.teaching.max_ones_problem.command_line": [[17, "module-opt.single_objective.teaching.max_ones_problem.command_line"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution"]], "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support": [[17, "module-opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support"]], "opt.single_objective.teaching.max_ones_problem.solver": [[17, "module-opt.single_objective.teaching.max_ones_problem.solver"]], "parse_arguments() (in module opt.single_objective.teaching.max_ones_problem.command_line)": [[17, "opt.single_objective.teaching.max_ones_problem.command_line.parse_arguments"]], "random() (in module opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.random"]], "random() (in module opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.random"]], "representation_distance_directly() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.representation_distance_directly"]], "representation_distance_directly() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.representation_distance_directly"]], "shaking() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.shaking"]], "shaking() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.shaking"]], "string_rep() (opt.single_objective.teaching.max_ones_problem.max_ones_problem.maxonesproblem method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem.MaxOnesProblem.string_rep"]], "string_rep() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.string_rep"]], "string_rep() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.maxonesproblembinarybitarraysolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport.string_rep"]], "string_rep() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.string_rep"]], "string_rep() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.maxonesproblembinaryintsolutionvnssupport method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport.string_rep"]], "string_representation() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.maxonesproblembinarybitarraysolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution.MaxOnesProblemBinaryBitArraySolution.string_representation"]], "string_representation() (opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.maxonesproblembinaryintsolution method)": [[17, "opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution.string_representation"]], "uo": [[21, "module-uo"]], "algorithm (class in uo.algorithm.algorithm)": [[22, "uo.algorithm.algorithm.Algorithm"]], "outputcontrol (class in uo.algorithm.output_control)": [[22, "uo.algorithm.output_control.OutputControl"]], "best_solution (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.best_solution"]], "copy() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.copy"]], "copy_to_best_solution() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.copy_to_best_solution"]], "evaluation (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.evaluation"]], "execution_ended (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.execution_ended"]], "execution_started (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.execution_started"]], "fields (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.fields"]], "fields_definitions (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.fields_definitions"]], "fields_headings (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.fields_headings"]], "init() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.init"]], "is_first_solution_better() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.is_first_solution_better"]], "iteration (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.iteration"]], "moments (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.moments"]], "name (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.name"]], "output_control (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.output_control"]], "output_file (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.output_file"]], "string_rep() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.string_rep"]], "string_rep() (uo.algorithm.output_control.outputcontrol method)": [[22, "uo.algorithm.output_control.OutputControl.string_rep"]], "target_problem (uo.algorithm.algorithm.algorithm property)": [[22, "uo.algorithm.algorithm.Algorithm.target_problem"]], "uo.algorithm": [[22, "module-uo.algorithm"]], "uo.algorithm.algorithm": [[22, "module-uo.algorithm.algorithm"]], "uo.algorithm.output_control": [[22, "module-uo.algorithm.output_control"]], "write_after_algorithm (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_after_algorithm"]], "write_after_evaluation (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_after_evaluation"]], "write_after_iteration (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_after_iteration"]], "write_after_step_in_iteration (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_after_step_in_iteration"]], "write_before_algorithm (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_before_algorithm"]], "write_before_evaluation (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_before_evaluation"]], "write_before_iteration (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_before_iteration"]], "write_before_step_in_iteration (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_before_step_in_iteration"]], "write_output_headers_if_needed() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.write_output_headers_if_needed"]], "write_output_values_if_needed() (uo.algorithm.algorithm.algorithm method)": [[22, "uo.algorithm.algorithm.Algorithm.write_output_values_if_needed"]], "write_to_output (uo.algorithm.output_control.outputcontrol property)": [[22, "uo.algorithm.output_control.OutputControl.write_to_output"]], "uo.algorithm.exact": [[23, "module-uo.algorithm.exact"]], "problemsolutiontesupport (class in uo.algorithm.exact.total_enumeration.problem_solution_te_support)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport"]], "teoptimizer (class in uo.algorithm.exact.total_enumeration.te_optimizer)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer"]], "teoptimizerconstructionparameters (class in uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters"]], "can_progress() (uo.algorithm.exact.total_enumeration.problem_solution_te_support.problemsolutiontesupport method)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport.can_progress"]], "copy() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.copy"]], "current_solution (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer property)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.current_solution"]], "from_construction_tuple() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer class method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.from_construction_tuple"]], "init() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.init"]], "initial_solution (uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.teoptimizerconstructionparameters attribute)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters.initial_solution"]], "iteration (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer property)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.iteration"]], "optimize() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.optimize"]], "output_control (uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.teoptimizerconstructionparameters attribute)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters.output_control"]], "overall_number_of_evaluations() (uo.algorithm.exact.total_enumeration.problem_solution_te_support.problemsolutiontesupport method)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport.overall_number_of_evaluations"]], "problem_solution_te_support (uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.teoptimizerconstructionparameters attribute)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters.problem_solution_te_support"]], "progress() (uo.algorithm.exact.total_enumeration.problem_solution_te_support.problemsolutiontesupport method)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport.progress"]], "random() (in module uo.algorithm.exact.total_enumeration.te_optimizer)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.random"]], "reset() (uo.algorithm.exact.total_enumeration.problem_solution_te_support.problemsolutiontesupport method)": [[24, "uo.algorithm.exact.total_enumeration.problem_solution_te_support.ProblemSolutionTeSupport.reset"]], "string_rep() (uo.algorithm.exact.total_enumeration.te_optimizer.teoptimizer method)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer.TeOptimizer.string_rep"]], "target_problem (uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.teoptimizerconstructionparameters attribute)": [[24, "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters.TeOptimizerConstructionParameters.target_problem"]], "uo.algorithm.exact.total_enumeration": [[24, "module-uo.algorithm.exact.total_enumeration"]], "uo.algorithm.exact.total_enumeration.problem_solution_te_support": [[24, "module-uo.algorithm.exact.total_enumeration.problem_solution_te_support"]], "uo.algorithm.exact.total_enumeration.te_optimizer": [[24, "module-uo.algorithm.exact.total_enumeration.te_optimizer"]], "uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters": [[24, "module-uo.algorithm.exact.total_enumeration.te_optimizer_constructor_parameters"]], "metaheuristic (class in uo.algorithm.metaheuristic.metaheuristic)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic"]], "additional_statistics_control (uo.algorithm.metaheuristic.metaheuristic.metaheuristic property)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.additional_statistics_control"]], "copy() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.copy"]], "elapsed_seconds() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.elapsed_seconds"]], "finish_control (uo.algorithm.metaheuristic.metaheuristic.metaheuristic property)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.finish_control"]], "main_loop() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.main_loop"]], "main_loop_iteration() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.main_loop_iteration"]], "optimize() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.optimize"]], "random() (in module uo.algorithm.metaheuristic.metaheuristic)": [[25, "uo.algorithm.metaheuristic.metaheuristic.random"]], "random_seed (uo.algorithm.metaheuristic.metaheuristic.metaheuristic property)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.random_seed"]], "string_rep() (uo.algorithm.metaheuristic.metaheuristic.metaheuristic method)": [[25, "uo.algorithm.metaheuristic.metaheuristic.Metaheuristic.string_rep"]], "uo.algorithm.metaheuristic": [[25, "module-uo.algorithm.metaheuristic"]], "uo.algorithm.metaheuristic.metaheuristic": [[25, "module-uo.algorithm.metaheuristic.metaheuristic"]], "uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic": [[26, "module-uo.algorithm.metaheuristic.electro_magnetism_like_metaheusristic"]], "uo.algorithm.metaheuristic.genetic_algorithm": [[27, "module-uo.algorithm.metaheuristic.genetic_algorithm"]], "problemsolutionvnssupport (class in uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport"]], "vnsoptimizer (class in uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer"]], "copy() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.copy"]], "from_construction_tuple() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer class method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.from_construction_tuple"]], "init() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.init"]], "k_max (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer property)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.k_max"]], "k_min (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer property)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.k_min"]], "local_search_best_improvement() (uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.problemsolutionvnssupport method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport.local_search_best_improvement"]], "local_search_first_improvement() (uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.problemsolutionvnssupport method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport.local_search_first_improvement"]], "main_loop_iteration() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.main_loop_iteration"]], "random() (in module uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.random"]], "shaking() (uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.problemsolutionvnssupport method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support.ProblemSolutionVnsSupport.shaking"]], "string_rep() (uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.vnsoptimizer method)": [[28, "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer.VnsOptimizer.string_rep"]], "uo.algorithm.metaheuristic.variable_neighborhood_search": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support"]], "uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer": [[28, "module-uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer"]], "targetproblem (class in uo.target_problem.target_problem)": [[29, "uo.target_problem.target_problem.TargetProblem"]], "copy() (uo.target_problem.target_problem.targetproblem method)": [[29, "uo.target_problem.target_problem.TargetProblem.copy"]], "dimension (uo.target_problem.target_problem.targetproblem property)": [[29, "uo.target_problem.target_problem.TargetProblem.dimension"]], "file_path (uo.target_problem.target_problem.targetproblem property)": [[29, "uo.target_problem.target_problem.TargetProblem.file_path"]], "is_minimization (uo.target_problem.target_problem.targetproblem property)": [[29, "uo.target_problem.target_problem.TargetProblem.is_minimization"]], "load_from_file() (uo.target_problem.target_problem.targetproblem method)": [[29, "uo.target_problem.target_problem.TargetProblem.load_from_file"]], "name (uo.target_problem.target_problem.targetproblem property)": [[29, "uo.target_problem.target_problem.TargetProblem.name"]], "string_rep() (uo.target_problem.target_problem.targetproblem method)": [[29, "uo.target_problem.target_problem.TargetProblem.string_rep"]], "uo.target_problem": [[29, "module-uo.target_problem"]], "uo.target_problem.target_problem": [[29, "module-uo.target_problem.target_problem"]], "evaluationcachecontrolstatistics (class in uo.target_solution.evaluation_cache_control_statistics)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics"]], "qualityofsolution (class in uo.target_solution.target_solution)": [[30, "uo.target_solution.target_solution.QualityOfSolution"]], "targetsolution (class in uo.target_solution.target_solution)": [[30, "uo.target_solution.target_solution.TargetSolution"]], "cache (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.cache"]], "cache_hit_count (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.cache_hit_count"]], "cache_request_count (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.cache_request_count"]], "calculate_quality() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.calculate_quality"]], "calculate_quality_directly() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.calculate_quality_directly"]], "copy() (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics method)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.copy"]], "copy() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.copy"]], "copy_to() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.copy_to"]], "evaluate() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.evaluate"]], "fitness_value (uo.target_solution.target_solution.qualityofsolution attribute)": [[30, "uo.target_solution.target_solution.QualityOfSolution.fitness_value"]], "fitness_value (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.fitness_value"]], "increment_cache_hit_count() (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics method)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.increment_cache_hit_count"]], "increment_cache_request_count() (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics method)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.increment_cache_request_count"]], "init_from() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.init_from"]], "init_random() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.init_random"]], "is_caching (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.is_caching"]], "is_feasible (uo.target_solution.target_solution.qualityofsolution attribute)": [[30, "uo.target_solution.target_solution.QualityOfSolution.is_feasible"]], "is_feasible (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.is_feasible"]], "max_cache_size (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics property)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.max_cache_size"]], "name (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.name"]], "native_representation() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.native_representation"]], "objective_value (uo.target_solution.target_solution.qualityofsolution attribute)": [[30, "uo.target_solution.target_solution.QualityOfSolution.objective_value"]], "objective_value (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.objective_value"]], "representation (uo.target_solution.target_solution.targetsolution property)": [[30, "uo.target_solution.target_solution.TargetSolution.representation"]], "representation_distance() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.representation_distance"]], "representation_distance_directly() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.representation_distance_directly"]], "string_rep() (uo.target_solution.evaluation_cache_control_statistics.evaluationcachecontrolstatistics method)": [[30, "uo.target_solution.evaluation_cache_control_statistics.EvaluationCacheControlStatistics.string_rep"]], "string_rep() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.string_rep"]], "string_representation() (uo.target_solution.target_solution.targetsolution method)": [[30, "uo.target_solution.target_solution.TargetSolution.string_representation"]], "uo.target_solution": [[30, "module-uo.target_solution"]], "uo.target_solution.evaluation_cache_control_statistics": [[30, "module-uo.target_solution.evaluation_cache_control_statistics"]], "uo.target_solution.target_solution": [[30, "module-uo.target_solution.target_solution"]], "ensure_dir() (in module uo.utils.files)": [[31, "uo.utils.files.ensure_dir"]], "uo.utils": [[31, "module-uo.utils"]], "uo.utils.files": [[31, "module-uo.utils.files"]], "uo.utils.logger": [[31, "module-uo.utils.logger"]]}})
