+++ documentation/source/opt.single_objective.trivial.rst
@@ -1,4 +1,4 @@
-opt.single\_objective.trivial package
+opt.single\_objective.teaching package
 =====================================
 
 Subpackages
@@ -7,12 +7,12 @@ Subpackages
 .. toctree::
    :maxdepth: 4
 
-   opt.single_objective.trivial.max_ones_problem
+   opt.single_objective.teaching.max_ones_problem
 
 Module contents
 ---------------
 
-.. automodule:: opt.single_objective.trivial
+.. automodule:: opt.single_objective.teaching
    :members:
    :undoc-members:
    :show-inheritance:
