+++ opt/single_objective/teaching/max_ones_problem/inputs/dimension_10.txt
