+++ opt/single_objective/teaching/max_ones_problem/max_ones_problem_binary_int_solution.py
@@ -1,7 +1,7 @@
 """ 
 .. _py_max_ones_problem_int_solution:
 
-The :mod:`~opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution` contains class :class:`~opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution`, that represents solution of the :ref:`Problem_Max_Ones`, where `int` representation of the problem has been used.
+The :mod:`~opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution` contains class :class:`~opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution.MaxOnesProblemBinaryIntSolution`, that represents solution of the :ref:`Problem_Max_Ones`, where `int` representation of the problem has been used.
 """
 
 import sys
