+++ opt/single_objective/teaching/max_ones_problem/max_ones_problem_binary_int_solution_vns_support.py
@@ -1,26 +1,23 @@
 """ 
-..  _py_max_ones_problem_bit_array_solution_vns_support:
+.. _py_max_ones_problem_int_solution_vns_support:
 
-The :mod:`~opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support` 
-contains class :class:`~opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support.MaxOnesProblemBinaryBitArraySolutionVnsSupport`, 
-that represents supporting parts of the `VNS` algorithm, where solution of the :ref:`Problem_Max_Ones` have `BitArray` 
-representation.
+The :mod:`~opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support` contains 
+class :class:`~opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution_vns_support.MaxOnesProblemBinaryIntSolutionVnsSupport`, 
+that represents solution of the :ref:`Problem_Max_Ones`, where `int` representation of the problem has been used.
 """
 
 import sys
 from pathlib import Path
 directory = Path(__file__).resolve()
-sys.path.append(directory)
 sys.path.append(directory.parent)
+sys.path.append(directory.parent.parent)
 sys.path.append(directory.parent.parent.parent)
 sys.path.append(directory.parent.parent.parent.parent)
 sys.path.append(directory.parent.parent.parent.parent.parent)
 
 from copy import deepcopy
 from random import choice
-from random import random
-
-from bitstring import Bits, BitArray, BitStream, pack
+from random import randint
 
 from uo.utils.logger import logger
 from uo.utils.complex_counter_uniform_distinct import ComplexCounterUniformAscending
@@ -29,51 +26,50 @@ from uo.target_solution.target_solution import QualityOfSolution
 from uo.algorithm.algorithm import Algorithm
 from uo.algorithm.metaheuristic.variable_neighborhood_search.problem_solution_vns_support import ProblemSolutionVnsSupport
 
-from opt.single_objective.trivial.max_ones_problem.max_ones_problem import MaxOnesProblem
-from opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_bit_array_solution import MaxOnesProblemBinaryBitArraySolution
+from opt.single_objective.teaching.max_ones_problem.max_ones_problem import MaxOnesProblem
+from opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_int_solution import MaxOnesProblemBinaryIntSolution
 
-class MaxOnesProblemBinaryBitArraySolutionVnsSupport(ProblemSolutionVnsSupport[BitArray]):
+class MaxOnesProblemBinaryIntSolutionVnsSupport(ProblemSolutionVnsSupport[int]):
     
     def __init__(self)->None:
         """
-        Create new `MaxOnesProblemBinaryBitArraySolutionVnsSupport` instance
+        Create new `MaxOnesProblemBinaryIntSolutionVnsSupport` instance
         """
         return
 
     def __copy__(self):
         """
-        Internal copy of the `MaxOnesProblemBinaryBitArraySolutionVnsSupport`
+        Internal copy of the `MaxOnesProblemBinaryIntSolutionVnsSupport`
 
-        :return: new `MaxOnesProblemBinaryBitArraySolutionVnsSupport` instance with the same properties
-        :rtype: `MaxOnesProblemBinaryBitArraySolutionVnsSupport`
+        :return: new `MaxOnesProblemBinaryIntSolutionVnsSupport` instance with the same properties
+        :rtype: MaxOnesProblemBinaryIntSolutionVnsSupport
         """
-        sol = deepcopy(self)
-        return sol
+        sup = deepcopy(self)
+        return sup
 
     def copy(self):
         """
-        Copy the `MaxOnesProblemBinaryBitArraySolutionVnsSupport` instance
-
-        :return: new `MaxOnesProblemBinaryBitArraySolutionVnsSupport` instance with the same properties
-        :rtype: `MaxOnesProblemBinaryBitArraySolutionVnsSupport`
-        """
+        Copy the `MaxOnesProblemBinaryIntSolutionVnsSupport`
+        
+        :return: new `MaxOnesProblemBinaryIntSolutionVnsSupport` instance with the same properties
+        :rtype: `MaxOnesProblemBinaryIntSolutionVnsSupport`
+        """        
         return self.__copy__()
-
-    def shaking(self, k:int, problem:MaxOnesProblem, solution:MaxOnesProblemBinaryBitArraySolution, 
+        
+    def shaking(self, k:int, problem:MaxOnesProblem, solution:MaxOnesProblemBinaryIntSolution, 
             optimizer:Algorithm)->bool:
         """
-        Random shaking of k parts such that new solution code does not differ more than k from all solution codes 
+        Random VNS shaking of k parts such that new solution code does not differ more than k from all solution codes 
         inside shakingPoints 
 
         :param int k: int parameter for VNS
         :param `MaxOnesProblem` problem: problem that is solved
-        :param `MaxOnesProblemBinaryBitArraySolution` solution: solution used for the problem that is solved
+        :param `MaxOnesProblemBinaryIntSolution` solution: solution used for the problem that is solved
         :param `Algorithm` optimizer: optimizer that is executed
-        :return: if randomization is successful
+        :return: if shaking is successful
         :rtype: bool
         """    
-        if optimizer.finish_control.evaluations_max > 0 \
-                and optimizer.evaluation > optimizer.finish_control.evaluations_max:
+        if optimizer.finish_control.evaluations_max > 0 and optimizer.evaluation > optimizer.finish_control.evaluations_max:
             return False
         tries:int = 0
         limit:int = 10000
@@ -81,65 +77,67 @@ class MaxOnesProblemBinaryBitArraySolutionVnsSupport(ProblemSolutionVnsSupport[B
             positions:list[int] = []
             for i in range(0,k):
                 positions.append(choice(range(problem.dimension)))
-            repr:BitArray = BitArray(solution.representation.tobytes())
-            for pos in positions:
-                repr[pos] = not repr[pos]
-            solution.representation = repr
+            mask:int = 0
+            for p in positions:
+                mask |= 1 << p
+            solution.representation ^= mask
             all_ok:bool = True
-            if solution.representation.count(value=1) > problem.dimension:
+            if solution.representation.bit_count() > problem.dimension:
                 all_ok = False
             if all_ok:
                 break
         if tries < limit:
             optimizer.evaluation += 1
             if optimizer.finish_control.evaluations_max > 0 and optimizer.evaluation > optimizer.finish_control.evaluations_max:
-                return False
+                return solution
             optimizer.write_output_values_if_needed("before_evaluation", "b_e")
             solution.evaluate(problem)
             optimizer.write_output_values_if_needed("after_evaluation", "a_e")
-            optimizer.write_output_values_if_needed("after_step_in_iteration", "shaking")
             return True
         else:
             return False 
 
-    def local_search_best_improvement(self, k:int, problem:MaxOnesProblem, solution:MaxOnesProblemBinaryBitArraySolution, 
-            optimizer: Algorithm)->MaxOnesProblemBinaryBitArraySolution:
+    def local_search_best_improvement(self, k:int, problem:MaxOnesProblem, solution:MaxOnesProblemBinaryIntSolution, 
+            optimizer: Algorithm)->MaxOnesProblemBinaryIntSolution:
         """
         Executes "best improvement" variant of the local search procedure 
         
         :param int k: int parameter for VNS
         :param `MaxOnesProblem` problem: problem that is solved
-        :param `MaxOnesProblemBinaryBitArraySolution` solution: solution used for the problem that is solved
+        :param `MaxOnesProblemBinaryIntSolution` solution: solution used for the problem that is solved
         :param `Algorithm` optimizer: optimizer that is executed
         :return: result of the local search procedure 
-        :rtype: MaxOnesProblemBinaryBitArraySolution
+        :rtype: MaxOnesProblemBinaryIntSolution
         """
         if optimizer.finish_control.evaluations_max > 0 and optimizer.evaluation > optimizer.finish_control.evaluations_max:
             return solution
         if k < 1 or k > problem.dimension:
             return solution
-        best_rep:BitArray = None
+        best_rep:int = None
         best_triplet:QualityOfSolution =  QualityOfSolution(solution.objective_value,
                 solution.fitness_value, solution.is_feasible)
         # initialize indexes
-        indexes:ComplexCounterUniformAscending = ComplexCounterUniformAscending(k, problem.dimension)
+        indexes:ComplexCounterUniformAscending = ComplexCounterUniformAscending(k,problem.dimension)
         in_loop:boolean = indexes.reset()
         while in_loop:
             # collect positions for inversion from indexes
             positions:list[int] = indexes.current_state()
             # invert and compare, switch of new is better
-            solution.representation.invert(positions) 
+            mask:int = 0
+            for i in positions:
+                mask |= 1 << i
+            solution.representation ^= mask 
             optimizer.evaluation += 1
             if optimizer.finish_control.evaluations_max > 0 and optimizer.evaluation > optimizer.finish_control.evaluations_max:
                 return solution
             optimizer.write_output_values_if_needed("before_evaluation", "b_e")
-            new_triplet:QualityOfSolution = solution.calculate_quality(problem)
+            new_triplet = solution.calculate_quality(problem)
             optimizer.write_output_values_if_needed("after_evaluation", "a_e")
             if new_triplet.fitness_value > best_triplet.fitness_value:
                 best_triplet = new_triplet
-                best_rep = BitArray(bin=solution.representation.bin)
-            solution.representation.invert(positions)
-            # increment indexes and set in_loop according to the state
+                best_rep = solution.representation
+            solution.representation ^= mask 
+            # increment indexes and set in_loop accordingly
             in_loop = indexes.progress()
         if best_rep is not None:
             solution.representation = best_rep
@@ -149,17 +147,17 @@ class MaxOnesProblemBinaryBitArraySolutionVnsSupport(ProblemSolutionVnsSupport[B
             return solution
         return solution
 
-    def local_search_first_improvement(self, k:int, problem:MaxOnesProblem, solution:MaxOnesProblemBinaryBitArraySolution, 
-            optimizer: Algorithm)->MaxOnesProblemBinaryBitArraySolution:
+    def local_search_first_improvement(self, k:int, problem:MaxOnesProblem, solution:MaxOnesProblemBinaryIntSolution, 
+            optimizer: Algorithm)->MaxOnesProblemBinaryIntSolution:
         """
         Executes "first improvement" variant of the local search procedure 
         
         :param int k: int parameter for VNS
         :param `MaxOnesProblem` problem: problem that is solved
-        :param `MaxOnesProblemBinaryBitArraySolution` solution: solution used for the problem that is solved
+        :param `MaxOnesProblemBinaryIntSolution` solution: solution used for the problem that is solved
         :param `Algorithm` optimizer: optimizer that is executed
         :return: result of the local search procedure 
-        :rtype: MaxOnesProblemBinaryBitArraySolution
+        :rtype: MaxOnesProblemBinaryIntSolution
         """
         if optimizer.finish_control.evaluations_max > 0 and optimizer.evaluation > optimizer.finish_control.evaluations_max:
             return solution
@@ -167,25 +165,28 @@ class MaxOnesProblemBinaryBitArraySolutionVnsSupport(ProblemSolutionVnsSupport[B
             return solution
         best_fv:float = solution.fitness_value
         # initialize indexes
-        indexes:ComplexCounterUniformAscending = ComplexCounterUniformAscending(k, problem.dimension)
+        indexes:ComplexCounterUniformAscending = ComplexCounterUniformAscending(k,problem.dimension)
         in_loop:boolean = indexes.reset()
         while in_loop:
             # collect positions for inversion from indexes
             positions:list[int] = indexes.current_state()
             # invert and compare, switch and exit if new is better
-            solution.representation.invert(positions) 
+            mask:int = 0
+            for i in positions:
+                mask |= 1 << i
+            solution.representation ^= mask 
             optimizer.evaluation += 1
             if optimizer.finish_control.evaluations_max > 0 and optimizer.evaluation > optimizer.finish_control.evaluations_max:
                 return solution
             optimizer.write_output_values_if_needed("before_evaluation", "b_e")
-            new_triplet:QualityOfSolution = solution.calculate_quality(problem)
+            new_triplet = solution.calculate_quality(problem)
             optimizer.write_output_values_if_needed("after_evaluation", "a_e")
             if new_triplet.fitness_value > best_fv:
-                solution.objective_value = new_triplet.objective_value
                 solution.fitness_value = new_triplet.fitness_value
+                solution.objective_value = new_triplet.objective_value
                 solution.is_feasible = new_triplet.is_feasible
                 return solution
-            solution.representation.invert(positions)
+            solution.representation ^= mask
             # increment indexes and set in_loop accordingly
             in_loop = indexes.progress()
         return solution
@@ -193,7 +194,7 @@ class MaxOnesProblemBinaryBitArraySolutionVnsSupport(ProblemSolutionVnsSupport[B
     def string_rep(self, delimiter:str, indentation:int=0, indentation_symbol:str='', group_start:str ='{', 
         group_end:str ='}')->str:
         """
-        String representation of the vns support structure
+        String representation of the vns support instance
 
         :param delimiter: delimiter between fields
         :type delimiter: str
@@ -208,7 +209,7 @@ class MaxOnesProblemBinaryBitArraySolutionVnsSupport(ProblemSolutionVnsSupport[B
         :return: string representation of vns support instance
         :rtype: str
         """        
-        return 'MaxOnesProblemBinaryBitArraySolutionVnsSupport'
+        return 'MaxOnesProblemBinaryIntSolutionVnsSupport'
 
     def __str__(self)->str:
         """
@@ -238,5 +239,3 @@ class MaxOnesProblemBinaryBitArraySolutionVnsSupport(ProblemSolutionVnsSupport[B
         :rtype: str
         """
         return self.string_rep('|')
-
-
