+++ tests/opt/single_objective/teaching/max_ones_problem/test_integration_max_ones_problem_vns_binary_bit_array_solution_lsfi.py
@@ -17,6 +17,8 @@ from copy import deepcopy
 from random import randint
 from random import choice
 
+from bitstring import Bits, BitArray, BitStream, pack
+
 from uo.algorithm.output_control import OutputControl
 from uo.algorithm.metaheuristic.finish_control import FinishControl
 from uo.algorithm.metaheuristic.additional_statistics_control import AdditionalStatisticsControl
@@ -24,22 +26,23 @@ from uo.algorithm.metaheuristic.additional_statistics_control import AdditionalS
 from uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer_constructor_parameters import VnsOptimizerConstructionParameters
 from uo.algorithm.metaheuristic.variable_neighborhood_search.vns_optimizer import VnsOptimizer
 
-from opt.single_objective.trivial.max_ones_problem.max_ones_problem import MaxOnesProblem
-from opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution import MaxOnesProblemBinaryIntSolution
-from opt.single_objective.trivial.max_ones_problem.max_ones_problem_binary_int_solution_vns_support import MaxOnesProblemBinaryIntSolutionVnsSupport
+from opt.single_objective.teaching.max_ones_problem.max_ones_problem import MaxOnesProblem
+from opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution import MaxOnesProblemBinaryBitArraySolution
+from opt.single_objective.teaching.max_ones_problem.max_ones_problem_binary_bit_array_solution_vns_support import MaxOnesProblemBinaryBitArraySolutionVnsSupport
 
-class TestIntegrationMaxOnesProblemVnsBinaryIntSolutionLsbi(unittest.TestCase):
+class TestIntegrationMaxOnesProblemVnsBinaryBitArraySolutionLsbi(unittest.TestCase):
     
     @classmethod
     def setUpClass(cls):
-        print("setUpClass TestIntegrationMaxOnesProblemVnsBinaryIntSolutionLsbi\n")
+        print("setUpClass TestIntegrationMaxOnesProblemVnsBinaryBitArraySolutionLsbi\n")
 
     def setUp(self):
         self.output_control = OutputControl(False)
-        self.problem_to_solve:MaxOnesProblem = MaxOnesProblem(dim=22)
-        self.solution:MaxOnesProblemBinaryIntSolution = MaxOnesProblemBinaryIntSolution()
+        self.problem_to_solve:MaxOnesProblem = MaxOnesProblem(dim=24)
+        self.solution:MaxOnesProblemBinaryBitArraySolution = MaxOnesProblemBinaryBitArraySolution(random_seed=43434343)
         self.finish_control:FinishControl = FinishControl(criteria='evaluations', evaluations_max=500)
-        self.vns_support:MaxOnesProblemBinaryIntSolutionVnsSupport = MaxOnesProblemBinaryIntSolutionVnsSupport()
+        self.vns_support:MaxOnesProblemBinaryBitArraySolutionVnsSupport = \
+                MaxOnesProblemBinaryBitArraySolutionVnsSupport()
         self.additional_stat = AdditionalStatisticsControl(keep='')
         vns_construction_params:VnsOptimizerConstructionParameters = VnsOptimizerConstructionParameters()
         vns_construction_params.output_control = self.output_control
@@ -57,18 +60,24 @@ class TestIntegrationMaxOnesProblemVnsBinaryIntSolutionLsbi(unittest.TestCase):
         return
     
     def test_best_solution_after_optimization_should_be_all_optimal(self):
-        result = int('0b1111111111111111111111', base=0)
-        self.assertEqual(self.optimizer.best_solution.representation, result)
+        result:str = '111111111111111111111111'
+        self.assertEqual(self.optimizer.best_solution.string_representation(), result)
+
+    def test_best_solution_after_optimization_should_be_optimal_2(self):
+        self.assertEqual(len(self.optimizer.best_solution.string_representation()), self.problem_to_solve.dimension)
 
     def test_best_solution_after_optimization_should_have_optimal_fitness(self):
         self.assertEqual(self.optimizer.best_solution.fitness_value, self.problem_to_solve.dimension)
+
+    def test_best_solution_after_optimization_should_have_optimal_objective_value(self):
+        self.assertEqual(self.optimizer.best_solution.objective_value, self.problem_to_solve.dimension)
 
     def tearDown(self):
         return
 
     @classmethod
     def tearDownClass(cls):
-        print("\ntearDownClass TestIntegrationMaxOnesProblemVnsBinaryIntSolutionLsbi")
+        print("\ntearDownClass TestIntegrationMaxOnesProblemVnsBinaryBitArraySolutionLsbi")
     
 if __name__ == '__main__':
     unittest.main()
